% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_raster.R
\name{acquisitionTime}
\alias{acquisitionTime}
\title{Extract acquisition times}
\usage{
acquisitionTime(x)
}
\arguments{
\item{x}{A ppiRaster object}
}
\value{
The datetime slot
}
\description{
Extract acquisition times
}
