% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp.R
\name{as.data.frame.vp}
\alias{as.data.frame.vp}
\title{Coerce Vertical Profile data to a Data Frame}
\usage{
\method{as.data.frame}{vp}(x, row.names = NULL, optional = FALSE,
  quantities = names(x$data), suntime = TRUE, geo = TRUE, elev = -0.268,
  lat = NULL, lon = NULL, ...)
}
\arguments{
\item{x}{object of class vp}

\item{row.names}{\code{NULL} or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{If \code{FALSE} then the names of the variables in the data frame are checked to ensure that they are syntactically valid variable names and are not duplicated.}

\item{quantities}{an optional character vector with the names of the quantities to include as columns in the data frame}

\item{suntime}{logical. When TRUE, adds sunrise/sunset and day/night information to each row}

\item{geo}{logical. When TRUE, adds latitude, longitude and antenna height of the radar to each row}

\item{elev}{sun elevation in degrees, see \link{suntime}.}

\item{lat}{radar latitude in decimal degrees. When set, overrides the latitude stored in \code{x} in \link{suntime} calculations}

\item{lon}{radar longitude in decimal degrees. When set, overrides the longitude stored in \code{x} in \link{suntime} calculations}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
an object of class data.frame
}
\description{
Converts a vertical profile to a Data Frame,
and optionally adds information on sunrise/sunset, day/night and derived quantities
like migration traffic rates.
}
\details{
Note that only the 'dens' quantity is thresholded by the radial velocity standard deviation \link{sd_vvp}.
Note that this is different from the default \link{plot.vp}, \link{plot.vpts} and \link{get_quantity.vp} functions, where
quantities "eta","dbz","ff","u","v","w","dd" are all thresholded by \link{sd_vvp}
}
\examples{
# load an example vertical profile time series object
data(VP)
# convert the object to a data.frame
df=as.data.frame(VP)
# do not compute sunrise/sunset information
df=as.data.frame(VP,suntime=FALSE)
# override the latitude/longitude information stored in the object
# when calculating sunrise / sunset
df=as.data.frame(VP,suntime=TRUE,lat=50,lon=4)
}
