% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{as.data.frame.vpts}
\alias{as.data.frame.vpts}
\title{Coerce Vertical Profile Time Series to a Data Frame}
\usage{
\method{as.data.frame}{vpts}(x, suntime = T, elev = -0.268,
  lat = x$attributes$where$lat, lon = x$attributes$where$lon)
}
\arguments{
\item{x}{object of class vpts}

\item{suntime}{logical. When TRUE, adds sunrise/sunset and day/night information}

\item{elev}{sun elevation in degrees}

\item{lat}{radar latitude in decimal degrees}

\item{lon}{radar longitude in decimal degrees}
}
\value{
an object of class data.frame
}
\description{
Converts vertical profile time series (objects of class \code{vpts}) to a Data Frame,
and optionally adds information on sunrise/sunset, day/night and derived quantities
like migration traffic rates.
}
\examples{
# load an example vertical profile time series object
data(VPTS)
# convert the object to a data.frame
df=as.data.frame(VPTS)
# do not compute sunrise/sunset information
df=as.data.frame(VPTS,suntime=FALSE)
# override the latitude/longitude information stored in the object
df=as.data.frame(VPTS,suntime=TRUE,lat=50,lon=4)
}
