% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_raster.R
\name{as.ppiRaster}
\alias{as.ppiRaster}
\title{Convert an old ppi to a ppiRaster}
\usage{
as.ppiRaster(x, ...)
}
\arguments{
\item{x}{A 'ppi' object}
}
\value{
A ppiRaster object
}
\description{
Convert an old ppi to a ppiRaster
}
