% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.vp.R
\name{as.vp}
\alias{as.vp}
\title{Convert a dataframe into a vp object}
\usage{
as.vp(data)
}
\arguments{
\item{data}{a dataframe created from a VPTS CSV file}
}
\value{
a bioRad vp object
}
\description{
Convert a dataframe into a vp object
}
\examples{
# load vp data as a data.frame:
df <- as.data.frame(example_vp)
# convert the data.frame to a vp object:
as.vp(df)
}
