% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beam.R
\name{beam_profile_overlap}
\alias{beam_profile_overlap}
\title{Calculate overlap between a vertical profile and the radiation coverage pattern}
\usage{
beam_profile_overlap(vol, vp, range, ylim = c(0, 4000), steps = 500,
  quantity = "dens", normalize = T)
}
\arguments{
\item{vol}{a polar volume of class pvol}

\item{vp}{a vertical profile of class vp}

\item{range}{the distance(s) from the radar for which to calculate the overlap}

\item{ylim}{altitude range in meter, given as a numeric vector of length two.}

\item{quantity}{profile quantity to use for the altitude distribution, one of 'dens' or 'eta'.}

\item{normalize}{Whether to normalize the radiation coverage pattern over the altitude range specified by ylim}

\item{step}{altitude grid size used for numeric integrations}
}
\value{
A data.frame with columns range and overlap. Overlap is calculated as the
Bhattacharyya coefficient (i.e. distribution overlap) between the (normalized) vertical profile vp
and the (normalized) radiation coverage pattern as calculated by \link{beam_profile}
}
\description{
Calculates the distribution overlap between a vertical profile
and the radiation coverage pattern
}
\details{
to be written
}
\examples{
to be written
}
