% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{beamheight}
\alias{beamheight}
\title{Radar beam height}
\usage{
beamheight(range, elev)
}
\arguments{
\item{range}{numeric. Range in km}

\item{elev}{numeric. Elevation in degrees}
}
\value{
numeric value. Beam height in km
}
\description{
Calculates the height of a radar beam as a function of elevation and range, assuming the beam
is emitted at surface level.
}
\details{
To account for refraction of the beam towards the earth's surface, an effective earth's radius of 4/3 * (true radius) is assumed.
}

