% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{beamwidth}
\alias{beamwidth}
\title{Radar beam width}
\usage{
beamwidth(range, angle = 1)
}
\arguments{
\item{range}{numeric. Range (distance from the radar antenna) in km}

\item{angle}{numeric. Beam angle in degrees}
}
\value{
numeric value. Beam width in m
}
\description{
Calculates the width of a radar beam as a function of range and beam angle
}

