% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{bind}
\alias{bind}
\alias{bind.vp}
\alias{bind.vplist}
\alias{bind.vpts}
\title{Bind vertical profiles into a time series object}
\usage{
bind(x, ...)

\method{bind}{vp}(...)

\method{bind}{vplist}(x)

\method{bind}{vpts}(..., attributes.from = 1)
}
\arguments{
\item{...}{objects of class \code{vp} or \code{vpts}}

\item{attributes.from}{which vpts object to copy attributes form (default: first)}
}
\value{
an object of class \code{vpts}, see \link[bioRad]{summary.vpts} for details
The profiles in the input \code{vpts} objects will be sorted in time in the output object.
}
\description{
Use this function to combine profiles or time series objects of the same radar
}
\section{Methods (by class)}{
\itemize{
\item \code{vp}: bind \code{vp} objects into a \code{vpts} object. If \code{vp} of multiple radars are provided, a list
is returned containing \code{vpts} time series objects for each radar.

\item \code{vplist}: bind \code{vplist} objects into a \code{vpts} object. If data of multiple radars is provided, a list
is returned containing \code{vpts} time series objects for each radar.

\item \code{vpts}: bind multiple time series of vertical profiles (\code{vpts} objects) into a single \code{vpts} object
}}

\examples{
# load the example vpts object
data(VPTS)
# split the vpts object into two separate time series, one containing profile 1-10, and a second combining profile 11-20:
vpts1=VPTS[1:10]
vpts2=VPTS[11:20]
# use bind to merge the two together:
vpts1and2=bind(vpts1,vpts2)
# verify that the binded objected now has 20 profiles, 10 from vpts1 and 10 from vpts2:
summary(vpts1and2)
}
