% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\docType{package}
\name{bioRad-package}
\alias{bioRad}
\alias{bioRad-package}
\title{bioRad: Biological analysis and visualization of weather radar data}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

bioRad aims to standardize methods for extracting and reporting 
    biological signals from weather radars. It provides functionality to access 
    low‐level radar data, process these data into meaningful biological 
    information on animal speeds and directions at different altitudes in the 
    atmosphere, visualize these biological extractions, and calculate further 
    summary statistics.
}
\details{
To get started, see:

\itemize{
  \item \href{https://doi.org/10.1111/ecog.04028}{Dokter et al. (2018)}: a
  paper describing the package.
  \item \href{https://adokter.github.io/bioRad/articles/bioRad.html}{bioRad
  vignette}: an introduction to bioRad's main functionalities.
  \item \href{https://adokter.github.io/bioRad/reference/index.html}{Function
  reference}: an overview of all bioRad functions.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adokter/bioRad}
  \item \url{https://adokter.github.io/bioRad}
  \item Report bugs at \url{https://github.com/adokter/bioRad/issues}
}

}
\author{
\strong{Maintainer}: Adriaan M. Dokter \email{amd427@cornell.edu} (0000-0001-6573-066X)

Authors:
\itemize{
  \item Peter Desmet \email{peter.desmet@inbo.be} (0000-0002-8442-8025)
  \item Stijn Van Hoey \email{stijn.vanhoey@inbo.be} (0000-0001-6413-3185)
}

Other contributors:
\itemize{
  \item Jurriaan Spaaks \email{j.spaaks@esciencecenter.nl} [contributor]
  \item Lourens Veen \email{l.veen@esciencecenter.nl} [contributor]
  \item Liesbeth Verlinden \email{l.verlinden@uva.nl} (0000-0003-1744-9325) [contributor]
  \item Hidde Leijnse \email{hidde.leijnse@knmi.nl} (0000-0001-7835-4480) [contributor]
}

}
\keyword{internal}
