% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp.R
\name{c.vp}
\alias{c.vp}
\title{Concatenate vertical profiles (\code{vp}) into a list of vertical profiles}
\usage{
\method{c}{vp}(...)
}
\arguments{
\item{...}{\code{vp} objects.}
}
\value{
A list of \code{vp} objects.
}
\description{
Concatenates vertical profiles (\code{vp}) into a list of vertical profiles
(\code{c(vp, vp, vp)}) and warns if they are not from a single radar.
}
\examples{
# concatenate vp objects into a list:
c(example_vp, example_vp)
}
\seealso{
Other profile manipulation functions:
\code{\link[=bind_into_vpts]{bind_into_vpts()}},
\code{\link[=clean_mixture]{clean_mixture()}},
\code{\link[=filter_precip]{filter_precip()}},
\code{\link[=filter_vpts]{filter_vpts()}},
\code{\link[=rcs<-]{rcs<-()}},
\code{\link[=regularize_vpts]{regularize_vpts()}},
\code{\link[=sd_vvp_threshold<-]{sd_vvp_threshold<-()}}
}
\concept{profile manipulation functions}
