% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_param.R
\name{calculate_param}
\alias{calculate_param}
\title{Calculate a new scan parameter}
\usage{
calculate_param(x, ...)
}
\arguments{
\item{x}{an object of class \code{pvol} or class \code{scan}}

\item{...}{an expression defining the new scan parameter in terms of existing scan parameters}
}
\description{
Calculates a new scan parameter from a combination of existing scan parameters. Useful
for calculating quantities that are defined in terms of other basic radar moments, like
depolarization ratio, or for applying clutter corrections (CCORH) to uncorrected reflectivity
moments (TH), as in CCORH + TH
}
\examples{
# locate example volume file:
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")
# load the file:
pvol <- read_pvolfile(pvolfile)
# add depolarization ratio (DR) as a scan parameter (see Kilambi 2018):
pvol <- calculate_param(pvol, DR=10*log10((ZDR+1-2*ZDR^0.5*RHOHV)/(ZDR+1+2*ZDR^0.5*RHOHV)))
}
\references{
\itemize{
  \item Kilambi, A., Fabry, F., and Meunier, V., 2018. A simple and effective method
  for separating meteorological from nonmeteorological targets using dual‐polarization
  data. Journal of Atmospheric and Oceanic Technology, 35, 1415–1424.
  \url{https://doi.org/10.1175/JTECH-D-17-0175.1}
}
}
