% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_night.R
\name{check_night}
\alias{check_night}
\alias{check_night.default}
\alias{check_night.vp}
\alias{check_night.vplist}
\alias{check_night.vpts}
\alias{check_night.pvol}
\title{Calculate whether it is night at a geographic location and time}
\usage{
check_night(x, ..., elev = -0.268)

\method{check_night}{default}(lon, lat, date, elev = -0.268)

\method{check_night}{vp}(x, elev = -0.268)

\method{check_night}{vplist}(x, elev = -0.268)

\method{check_night}{vpts}(x, elev = -0.268)

\method{check_night}{pvol}(x, elev = -0.268)
}
\arguments{
\item{x}{An object of class \code{numeric}, \code{vp},
\code{vplist} or \code{vpts}.}

\item{elev}{sun elevation in degrees}

\item{lon}{longitude in decimal degrees}

\item{lat}{latitude in decimal degrees}

\item{date}{date inhereting from class "\code{POSIXt}" or a string
interpretable by \link[base]{as.Date}}
}
\value{
TRUE when night, FALSE when day, NA if unknown (either datetime or
geographic location missing). For objects of class vpts an atomic logical
vector
}
\description{
The night check is based on the combination of lat, lon, date and the sun
elevation. When testing a bioRad object for day/night time, this information
is extracted from the bioRad data types directly
}
\details{
The angular diameter of the sun is about 0.536 degrees, therefore
the moment of sunrise/sunset corresponds to half that elevation at
-0.268 degrees.

check_night evaluates to FALSE when the sun has a higher elevation than
parameter elev, otherwise to TRUE

Approximate astronomical formula are used, therefore the day/night
transition may be off by a few minutes
}
\examples{
# it's day in the Netherlands at UTC noon on January first:
check_night(5,53,"2016-01-01 12:00")
# applied on the bioRad object classes
check_night(VP)
check_night(VPTS)
}
