% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vpfiles.R
\name{check_radar_codes}
\alias{check_radar_codes}
\title{Check if radar codes are exactly 5 characters}
\usage{
check_radar_codes(radars)
}
\arguments{
\item{radars}{character vector. Radar codes to check, e.g. \code{c("bejab", "bewideu")}.}
}
\value{
NULL. Will stop and show error message if at least one of the
provided radar codes is not exactly 5 characters.
}
\description{
Check if radar codes are exactly 5 characters
}
\keyword{internal}
