% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{checkDocker}
\alias{checkDocker}
\title{Checks that Docker is running}
\usage{
checkDocker()
}
\description{
Checks that \href{https://www.docker.com/}{Docker} daemon is running correctly on the local system
}

