% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite.R
\name{composite}
\alias{composite}
\title{Make a composite of multiple plan position indicators (ppi objects)}
\usage{
composite(x, param = "DBZH", cells.dim = c(100, 100))
}
\arguments{
\item{x}{a list of objects of class 'ppi'}

\item{param}{scan parameter to composite}

\item{cells.dim}{integer; vector with number of cells in each spatial dimension}
}
\value{
an object of class '\link[=summary.ppi]{ppi}'.
}
\description{
Merge multiple plan position indicators (ppi objects). Can be used to make a composite of ppi's from multiple radars
}
\details{
The returned PPI is in WGS84 projection (longitude, latitude)
}
\examples{
# load a polar scan example object
data(SCAN)
# to be written ...
}
