% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{day}
\alias{day}
\alias{day.vp}
\alias{day.vplist}
\alias{day.vpts}
\alias{day.pvol}
\title{Test a bioRad object for night time}
\usage{
day(x, elev = -0.268)

\method{day}{vp}(x, elev = -0.268)

\method{day}{vplist}(x, elev = -0.268)

\method{day}{vpts}(x, elev = -0.268)

\method{day}{pvol}(x, elev = -0.268)
}
\arguments{
\item{x}{An object of class \code{vp},\code{vplist} or \code{vpts}.}

\item{elev}{sun elevation in degrees}
}
\value{
TRUE when night, FALSE when day, NA if unknown (either datetime or geographic location missing). For objects of class vpts an atomic logical vector
}
\description{
Test a bioRad object for day time. Dispatches to the logical inverse of \link{night}.
}
\examples{
day(VP)
}
