% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dbz_eta.R
\name{dbz2eta}
\alias{dbz2eta}
\title{Convert reflectivity factor to reflectivity}
\usage{
dbz2eta(dbz, wavelength, Km = 0.93)
}
\arguments{
\item{dbz}{reflectivity factor in dBZ}

\item{wavelength}{radar wavelength in cm}

\item{Km}{refractive index of water}
}
\value{
reflectivity in cm^2/km^3
}
\description{
Convert reflectivity factor to reflectivity
}
