% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\docType{data}
\name{docker}
\alias{docker}
\title{flag indicating whether docker is running}
\format{An object of class \code{logical} of length 1.}
\usage{
docker
}
\description{
flag indicating whether docker is running
}
\keyword{internal}

