% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vpts_aloft.R
\name{download_vpts_aloft}
\alias{download_vpts_aloft}
\title{Download vertical profile time series data (\code{vpts}) from aloftdata.eu}
\usage{
download_vpts_aloft(
  directory = ".",
  radars = NULL,
  start_date = NULL,
  end_date = NULL,
  source = "baltrad",
  format = "csv",
  overwrite = FALSE
)
}
\arguments{
\item{directory}{description}
}
\value{
invisibly return directory
}
\description{
Download vertical profile time series data (\code{vpts}) from aloftdata.eu
}
\examples{
# example code
}
