% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{elev}
\alias{elev}
\alias{elev.PVOL}
\alias{elev.SCAN}
\title{Elevation}
\usage{
elev(x, ...)

\method{elev}{SCAN}(x)

\method{elev}{PVOL}(x)
}
\arguments{
\item{x}{a \code{PVOL} or \code{SCAN} object}
}
\value{
elevation in degrees
}
\description{
Gives the elevation of a scan, or the elevations within a polar volume
}
\section{Methods (by class)}{
\itemize{
\item \code{SCAN}: elevation of a scan

\item \code{PVOL}: elevation of a polar volume
}}
\examples{
# to be written
}

