% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_scan}
\alias{example_scan}
\title{Scan (\code{scan}) example}
\format{
An object of class \code{scan} of dimension 5 x 480 x 360.
}
\usage{
example_scan
}
\description{
Example object of class \code{\link[=summary.scan]{scan}} with name \code{example_scan}.
}
\examples{
# Reload example_scan from package (e.g. in case it was altered)
data(example_scan)

# Get summary info
example_scan

# example_scan was created with
\dontrun{
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")
pvol <- read_pvolfile(pvolfile)
example_scan <- pvol$scans[[1]]
save(example_scan, file = "data/example_scan.rda")
}
}
\seealso{
\itemize{
\item \code{\link[=summary.scan]{summary.scan()}}
}
}
\keyword{datasets}
