% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_vp}
\alias{example_vp}
\title{Example object of class \code{vp}}
\format{
An object of class \code{vp} with 25 rows and 16 columns.
}
\usage{
example_vp
}
\description{
Example of a \code{\link[=summary.vp]{vp}} object with name
\code{example_vp}. Can be created with \code{\link{calculate_vp}} or read
from file with \code{\link{read_vpfiles}}.
}
\examples{
# get summary of example vp:
summary(example_vp)

# example_vp was created with:
vpfile <- system.file("extdata", "profile.h5", package = "bioRad")
example_vp <- read_vpfiles(vpfile)
\dontrun{
# save(example_vp, file = "data/example_vp.rda")
}
}
\keyword{datasets}
