% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_vpts}
\alias{example_vpts}
\title{Time series of vertical profiles (\code{vpts}) example}
\format{
An object of class \code{vpts} of dimension 1934 x 25 x 15.
}
\usage{
example_vpts
}
\value{
An example object of type 'vpts' which represents a time series of vertical profiles.
}
\description{
Example of a \code{\link[=summary.vpts]{vpts}} object with name \code{example_vpts}.
}
\examples{
# Reload example_vpts from package (e.g. in case it was altered)
data(example_vpts)

# Get summary info
example_vpts
}
\seealso{
\itemize{
\item \code{\link[=summary.vpts]{summary.vpts()}}
}
}
\keyword{datasets}
