% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{extract_string}
\alias{extract_string}
\title{extract strings from a vector using regex, analog to stringr::str_extract}
\usage{
extract_string(string, pattern, ...)
}
\arguments{
\item{string}{Input vector. A character vector.}

\item{pattern}{Regex pattern to look for}

\item{...}{passed on to \code{regexpr()}}
}
\value{
A character vector with matches only, possibly of different length as
\code{string}
}
\description{
extract strings from a vector using regex, analog to stringr::str_extract
}
\keyword{internal}
