% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_datetime}
\alias{filter_datetime}
\title{Time selection in a time series of vertical profiles ('vpts')}
\usage{
filter_datetime(x, min, max)
}
\arguments{
\item{x}{A \code{vpts} object.}

\item{min}{POSIXct value. Minimum datetime to be included}

\item{max}{POSIXct value. Maximum datetime to be included}
}
\value{
An object of class '\link[=summary.vpts]{vpts}'.
}
\description{
Time selection in a time series of vertical profiles ('vpts')
}
\details{
returns profiles for which min <= timestamp profile < max.
}
\examples{
# let us use this example vertical profile time series:
example_vpts
# select profiles later than 02-Sep-2016
filter_datetime(example_vpts, min=as.POSIXct("2016-09-02"))
}
