% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_as_corrected_ppi.R
\name{georeference_scan}
\alias{georeference_scan}
\title{project a scan into WGS84 (lat,lon) coordinates}
\usage{
georeference_scan(scan)
}
\arguments{
\item{scan}{a scan (sweep) of class scan}
}
\value{
a SpatialPointsDataFrame
TODO XXX BEAM CURVATURE NOT ACCOUNTED FOR when calculating range_ground_level ...
}
\description{
project the pixels of a scan in polar coordinates into WGS84 (lat,lon) coordinates
}
\keyword{internal}
