% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_angles.R
\name{get_angles}
\alias{get_angles}
\alias{get_angles.scan}
\alias{get_angles.pvol}
\title{Get elevation angle of a scan (\code{scan}) or scans in a polar volume (\code{scan})}
\usage{
get_angles(x)

\method{get_angles}{scan}(x)

\method{get_angles}{pvol}(x)
}
\arguments{
\item{x}{A \code{pvol} or \code{scan} object.}
}
\value{
elevation in degrees
}
\description{
Gives the elevation angle of a scan, or the elevation angles within a
polar volume
}
\section{Methods (by class)}{
\itemize{
\item \code{scan}: Elevation angle of a scan.

\item \code{pvol}: Elevation angles of all scans in a polar volume.
}}

\examples{
# load a polar volume
pvol <- system.file("extdata", "volume.h5", package="bioRad")
vol <- read_pvolfile(pvol)
# elevations for the scans in the volume
get_angles(vol)
# extract the first scan:
scan <- vol$scans[[1]]
# elevation angle of the scan:
get_angles(scan)
}
