% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevation_angles.R
\name{get_elevation_angles}
\alias{get_elevation_angles}
\alias{get_elevation_angles.pvol}
\alias{get_elevation_angles.scan}
\alias{get_elevation_angles.param}
\title{Get elevation angles of a polar volume (\code{pvol}) or scan (\code{scan})}
\usage{
get_elevation_angles(x)

\method{get_elevation_angles}{pvol}(x)

\method{get_elevation_angles}{scan}(x)

\method{get_elevation_angles}{param}(x)
}
\arguments{
\item{x}{A \code{pvol}, \code{scan} or \verb{scan parameter}.}
}
\value{
Elevation in degrees.
}
\description{
Gives the elevation angle of a single scan, or the elevation angles of all scans within a
polar volume
}
\examples{
# load a polar volume
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")
example_pvol <- read_pvolfile(pvolfile)

# elevations for the scans in the volume
get_elevation_angles(example_pvol)

# extract the first scan:
scan <- example_pvol$scans[[1]]

# elevation angle of the scan:
get_elevation_angles(scan)
}
