% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris.R
\name{get_iris_raw_task}
\alias{get_iris_raw_task}
\title{Check task of IRIS RAW file}
\usage{
get_iris_raw_task(
  file,
  header_size = 50,
  task = c("WIND", "SURVEILLANCE", "VOL_A", "VOL_B")
)
}
\arguments{
\item{file}{A string containing a file name.}

\item{header_size}{Number of header bytes to search}

\item{task}{task names to search for in the file header}
}
\value{
one of the \code{task} names found in the header, \code{NA} if none of the task names were found.
}
\description{
Checks which task (polar volume type) is contained in a IRIS RAW file
}
