% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunrise_sunset.R
\name{get_suntime}
\alias{get_suntime}
\title{Helper function to calculate sunrise or sunset}
\usage{
get_suntime(lon, lat, date, sign_angle, elev = -0.268)
}
\arguments{
\item{lon}{Longitude in decimal degrees.}

\item{lat}{Latitude in decimal degrees.}

\item{date}{Date inhereting from class \code{POSIXt} or a string
interpretable by \link[base]{as.Date}.}

\item{sign_angle}{Whether to output for rising (\code{1}) or setting
(\code{-1}) sun.}

\item{elev}{Sun elevation in degrees.}
}
\description{
Use the \code{\link{sunrise}} or \code{\link{sunset}} functions.
}
