% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.ppi.R
\name{getscan}
\alias{getscan}
\title{Extract a scan from a polar volume}
\usage{
getscan(x, angle)
}
\arguments{
\item{x}{an object of class 'pvol'}

\item{angle}{elevation angle}
}
\value{
an object of class '\link[=summary.scan]{scan}'.
}
\description{
Extract a scan from a polar volume
}
\details{
The function returns the scan with elevation angle closest to \code{angle}
}
\examples{
# locate example volume file:
pvol <- system.file("extdata", "volume.h5", package="bioRad")
# load the file:
vol=read.pvol(pvol)
# extract the scan at 3 degree elevation:
myscan = getscan(vol,3)
}
