% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_file_type}
\alias{guess_file_type}
\title{Guess the file type of a file}
\usage{
guess_file_type(file_path, n_lines = 5)
}
\arguments{
\item{file_path}{A character string containing the path to the file}

\item{n_lines}{An integer, the number of lines to read for guessing a CSV file}
}
\value{
A character string representing the guessed file type ("h5", "gz", "csv", or "txt")
}
\description{
Guess the file type of a file based on the first few
bytes (for HDF5 and gzip files) or lines (for CSV files). If no known file
type can be guessed, it assumes the file is a text file.
}
\keyword{internal}
