% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvolfile.R
\name{h5ODIMobject}
\alias{h5ODIMobject}
\title{Check the ODIM data class of a polar volume file}
\usage{
h5ODIMobject(filename)
}
\arguments{
\item{filename}{A string containing a filename}
}
\value{
character string "\code{pvol}" for polar volume, "\code{vp}" for vertical profile, otherwise \code{NA}
}
\description{
Checks which data class is contained in ODIM HDF5 file
}
\examples{
# locate a polar volume file
pvol <- system.file("extdata", "volume.h5", package="bioRad")
h5ODIMobject(pvol)   #> "pvol"

}
