% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_profile.R
\name{integrate_profile}
\alias{integrate_profile}
\alias{integrate_profile.vp}
\alias{integrate_profile.list}
\alias{integrate_profile.vpts}
\title{Vertically integrate profiles (\code{vp} or \code{vpts}) into an
integrated profile (\code{vpi})}
\usage{
integrate_profile(
  x,
  alt_min,
  alt_max,
  alpha = NA,
  interval_max = 3600,
  interval_replace = NA,
  height_quantile = NA
)

\method{integrate_profile}{vp}(
  x,
  alt_min = 0,
  alt_max = Inf,
  alpha = NA,
  interval_max = 3600,
  interval_replace = NA,
  height_quantile = NA
)

\method{integrate_profile}{list}(
  x,
  alt_min = 0,
  alt_max = Inf,
  alpha = NA,
  interval_max = 3600,
  interval_replace = NA,
  height_quantile = NA
)

\method{integrate_profile}{vpts}(
  x,
  alt_min = 0,
  alt_max = Inf,
  alpha = NA,
  interval_max = 3600,
  interval_replace = NA,
  height_quantile = NA
)
}
\arguments{
\item{x}{A \code{vp} or \code{vpts} object.}

\item{alt_min}{Numeric. Minimum altitude, in m.}

\item{alt_max}{Numeric. Maximum altitude, in m.}

\item{alpha}{Numeric. Migratory direction, in clockwise degrees from north.}

\item{interval_max}{Numeric. Maximum time interval belonging to a single
profile, in seconds. Traffic rates are set to zero at times \code{t} for which
no profiles can be found within the period \code{t - interval_max/2} to \code{t + interval_max/2}. Ignored for single profiles of class \code{vp}.}
}
\value{
A \code{vpi} object: a data frame with vertically integrated profile
quantities.
}
\description{
Performs a vertical integration of density, reflectivity and migration
traffic rate, and a vertical averaging of ground speed and direction weighted
by density.
}
\details{

}
\section{Methods (by class)}{
\itemize{
\item \code{integrate_profile(vp)}: Vertically integrate a vertical profile (\code{vp}).

\item \code{integrate_profile(list)}: Vertically integrate a list of vertical
profiles (\code{vp}).

\item \code{integrate_profile(vpts)}: Vertically integrate a time series of
vertical profiles (\code{vpts}).

}}
\examples{
# Calculate migration traffic rates for a single vp
integrate_profile(example_vp)

# Calculate migration traffic rates for a list of vps
integrate_profile(c(example_vp, example_vp))

# Calculate migration traffic rates for a vpts
vpi <- integrate_profile(example_vpts)

# Plot migration traffic rate (mtr) for the full air column
plot(integrate_profile(example_vpts))

# Plot migration traffic rate (mtr) for altitudes > 1 km above sea level
plot(integrate_profile(example_vpts, alt_min = 1000))

# Plot cumulative migration traffic rates (mt)
plot(integrate_profile(example_vpts), quantity = "mt")
# calculate median flight altitude (instead of default mean)
integrate_profile(example_vp, height_quantile=.5)
# calculate the 90\% percentile of the flight altitude distribution
integrate_profile(example_vpts, height_quantile=.9)
}
