% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpfile.R
\name{is.vpfile}
\alias{is.vpfile}
\title{Check if a file is a vertical profile (\code{vp})}
\usage{
is.vpfile(file)
}
\arguments{
\item{file}{Character. Path of the file to check.}
}
\value{
\code{TRUE} for a vertical profile file in readable format, otherwise
\code{FALSE}.
}
\description{
Checks whether a file is a vertical profile (\code{vp}) in the ODIM HDF5 format
that can be read with bioRad.
}
\examples{
# Locate the vertical profile example file
vpfile <- system.file("extdata", "profile.h5", package = "bioRad")

# Check if it is a vpfile
is.vpfile(vpfile)
}
\seealso{
Other profile read functions:
\code{\link[=download_vpfiles]{download_vpfiles()}},
\code{\link[=list_vpts_aloft]{list_vpts_aloft()}},
\code{\link[=read_cajun]{read_cajun()}},
\code{\link[=read_vpfiles]{read_vpfiles()}},
\code{\link[=read_vpts]{read_vpts()}},
\code{\link[=select_vpfiles]{select_vpfiles()}}
}
\concept{profile read functions}
