% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_filenames}
\alias{match_filenames}
\title{Match a set of regular expressions to a list of files}
\usage{
match_filenames(file_list, regex_list)
}
\arguments{
\item{file_list}{character vector. Haystack of filenames/filepaths.}

\item{regex_list}{character vector. Needle of regular expressions to which
filenames should comply.}
}
\value{
character vector. Subset of filenames from the file_list that comply
to the provided regular expressions in regex_list.
}
\description{
Match a set of regular expressions to a list of files and return those
filenames that comply to any of the provided regular expressions. This
function basically wraps a grep to make it work on vectors by combining the
vector of regex options as possible options.
}
\keyword{internal}
