% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{min_package_version}
\alias{min_package_version}
\title{Identify minimum required package version of dependencies.}
\usage{
min_package_version(pkg)
}
\arguments{
\item{pkg}{A character string with a package name.}
}
\value{
A character string with the numeric version. When
no version is specified or the package is not listed in Suggests
or Depends a value \code{NULL} is returned.
}
\description{
Identify the minimum required package version of dependencies
listed in Suggests or Imports of the DESCRIPTION file.
}
\keyword{internal}
