% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\docType{data}
\name{mounted}
\alias{mounted}
\title{flag indicating whether vol2bird docker container is mounted}
\format{An object of class \code{logical} of length 1.}
\usage{
mounted
}
\description{
flag indicating whether vol2bird docker container is mounted
}
\keyword{internal}

