% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{mt}
\alias{mt}
\title{Migration traffic}
\usage{
mt(x, alt.min = 0, alt.max = Inf)
}
\arguments{
\item{x}{a \code{VP}, \code{VPList} or \code{VPTimeSeries} object}

\item{alt.min}{minimum altitude in m}

\item{alt.max}{maximum altitude in m}
}
\value{
a numeric value equal to migration traffic in number of individuals / km
}
\description{
Total migration traffic, which is calculated by time-integration
of migration traffic rates. Migration traffic gives the number of individuals
that have passed per km perpendicular to the migratory direction at the
position of the radar for the full period of the time series
within the specified altitude band.
}

