% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{mtr}
\alias{mtr}
\alias{mtr.VP}
\alias{mtr.VPList}
\alias{mtr.VPTimeSeries}
\title{Migration traffic rate}
\usage{
mtr(x, ...)

\method{mtr}{VP}(x, alt.min = 0, alt.max = Inf)

\method{mtr}{VPList}(x, alt.min = 0, alt.max = Inf)

\method{mtr}{VPTimeSeries}(x, alt.min = 0, alt.max = Inf)
}
\arguments{
\item{x}{a \code{VP}, \code{VPList} or \code{VPTimeSeries} object}

\item{alt.min}{minimum altitude in m}

\item{alt.max}{maximum altitude in m}
}
\value{
an atomic vector with MTRs in individuals/km/h
}
\description{
Migration traffic rate (MTR) for an altitude layer, defined as the
number of targets crossing a 1 km line perpendicular to the migratory movement per hour
}
\section{Methods (by class)}{
\itemize{
\item \code{VP}: MTR of a vertical profile

\item \code{VPList}: MTR of a list of vertical profiles

\item \code{VPTimeSeries}: MTR of a time series of vertical profiles
}}
\examples{
### MTR for a single vertical profile ###
mtr(VP)

### MTRs for a list of vertical profiles ###
mtr(c(VP,VP))

### MTRs for a time series of vertical profiles ###
# locate example file:
VPtable <- system.file("extdata", "VPtable.txt", package="bioRad")
# load time series:
ts=readVP.table(VPtable,radar="KBGM", wavelength='S')
# print migration traffic rates:
mtr(ts)
# plot migration traffic rates for the full air column:
plot(ts$dates,mtr(ts),type='l',xlab="time [UTC]",ylab="MTR [birds/km/h]")
#' plot migration traffic rates for altitudes > 1 km above sea level
plot(ts$dates,mtr(ts,alt.min=1000),type='l',xlab="time [UTC]",ylab="MTR [birds/km/h]")
}

