% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexrad_odim.R
\name{nexrad_to_odim}
\alias{nexrad_to_odim}
\title{Convert RSL polar volume to ODIM hdf5 format}
\usage{
nexrad_to_odim(vol.in, vol.out, verbose = FALSE, mount = dirname(vol.in))
}
\arguments{
\item{vol.in}{Polar volume input file in RSL format.}

\item{vol.out}{Filename for the polar volume in ODIM hdf5 format to be
generated}

\item{verbose}{logical. When TRUE, pipe Docker stdout to R console. On
Windows always TRUE.}

\item{mount}{character. String with the mount point (a directory path) for
the Docker container.}
}
\value{
\code{TRUE} on success
}
\description{
Convert RSL polar volume to ODIM hdf5 format
}
