% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyquist_velocity.R
\name{nyquist_velocity}
\alias{nyquist_velocity}
\title{Nyquist velocity for a given pulse repetition frequency (PRF)}
\usage{
nyquist_velocity(wavelength, prf1, prf2 = NA)
}
\value{
Nyquist velocity in m/s.
}
\description{
Calculates the Nyquist velocity for a given pulse repetition frequency (PRF)
and wavelength of the radar. When specifying two PRFs, the extended Nyquist
velocity is given for a radar using the dual-PRF technique.
}
\examples{
nyquist_velocity(5.3, 2000, 1500)
}
