% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vpi.R
\name{plot.vpi}
\alias{plot.vpi}
\title{Plot an integrated profile (\code{vpi})}
\usage{
\method{plot}{vpi}(x, quantity = "mtr", xlab = "time",
  ylab = "migration traffic rate [#/km/h]", main = "MTR",
  night_shade = TRUE, nightshade = TRUE, elev = -0.268, lat = NULL,
  lon = NULL, ylim = NULL, ...)
}
\arguments{
\item{x}{1 class object inheriting from class \code{vpi}, typically a
call to \link[bioRad]{integrate_profile}.}

\item{quantity}{Character string with the quantity to plot, one of
'\code{vid}','\code{vir}','\code{mtr}' for vertically integrated density,
reflectivity, reflectivity factor and migration traffic rate, respectively.}

\item{xlab}{A title for the x-axis.}

\item{ylab}{A title for the y-axis.}

\item{main}{A title for the plot.}

\item{night_shade}{Logical, whether to plot night time shading.}

\item{nightshade}{Deprecated argument, use night_shade instead.}

\item{elev}{Numeric, sun elevation to use for day/night transition,
see \link[bioRad]{suntime}.}

\item{lat}{(optional) Latitude in decimal degrees. Overrides the lat
attribute of \code{x}.}

\item{lon}{(optional) Longitude in decimal degrees. Overrides the lon
attribute of \code{x}.}

\item{ylim}{y-axis plot range, numeric atomic vector of length 2.}

\item{...}{Additional arguments to be passed to the low level
\link[graphics]{plot} plotting function.}
}
\description{
Plot an object of class \code{vpi}.
}
\details{
Profile can be visualised in three related quantities, as specified by
argument \code{quantity}:
\describe{
 \item{"\code{vid}"}{Vertically Integrated Density, i.e. the aerial surface
   density of individuals. This quantity is dependent on the assumed radar
   cross section per individual (RCS)}
 \item{"\code{vir}"}{Vertically Integrated Reflectivity. This quantity is
   independent of the value of individual's radar cross section}
 \item{"\code{mtr}"}{Migration Traffic Rate. This quantity is dependent on
   the assumed radar cross section (RCS)}
 \item{"\code{rtr}"}{Reflectivity Traffic Rate. This quantity is independent
   on the assumed radar cross section (RCS)}
}
}
\examples{
# vertically integrate a vpts object:
vpi <- integrate_profile(example_vpts)
# plot the migration traffic rates
plot(vpi)
# plot the vertically integrated densities, without night shading:
plot(vpi, quantity = "vid", night_shade = FALSE)
}
