% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_raster.R, R/plot.ppi.raster.R
\docType{class}
\name{ppiRaster-class}
\alias{ppiRaster-class}
\alias{ppi_to_ppiRaster}
\alias{show,ppiRaster-method}
\alias{plot,ppiRaster,missing-method}
\alias{plot,ppiRaster,missing-method,plot.ppiRaster}
\title{ppiRaster: S4 subclass of terra::SpatRaster}
\usage{
ppi_to_ppiRaster(ppi)

\S4method{show}{ppiRaster}(object)

\S4method{plot}{ppiRaster,missing}(x, y, param, zlim, xlim, ylim, ratio = 1, na.value = "transparent", ...)
}
\arguments{
\item{ppi}{A list of class 'ppi' as returned by bioRad::project_as_ppi()}

\item{object}{An object of class ppiRaster}

\item{x}{A ppiRaster object}

\item{y}{Missing (for S4 method signature)}

\item{param}{Character. Layer name to plot (default first layer)}

\item{zlim}{Numeric\link{2}. Value range (default full data range)}

\item{xlim, ylim}{Numeric\link{2}. Spatial extents (default raster extent)}

\item{ratio}{Numeric. Aspect ratio y/x (default 1)}

\item{na.value}{Color for NA values (default "transparent")}

\item{...}{Additional arguments passed to ggplot2::ggplot}
}
\value{
An object of class ppiRaster
}
\description{
Extends terra::SpatRaster by embedding radar metadata
dedicated S4 slots, with formal validity checks and custom printing.
}
