% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.R
\name{print.ppi}
\alias{print.ppi}
\title{Print method for ppi}
\usage{
\method{print}{ppi}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{ppi}.}
}
\description{
Print method for ppi
}
\keyword{internal}
