% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{print.VPList}
\alias{print.VPList}
\title{print method for class \code{VPList}}
\usage{
\method{print}{VPList}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{VPList}, usually a result of a call to \link[bioRad]{readVP.list}}
}
\description{
print method for class \code{VPList}
}
\keyword{internal}

