% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{print.VPTimeSeries}
\alias{print.VPTimeSeries}
\title{print method for class \code{VPTimeSeries}}
\usage{
\method{print}{VPTimeSeries}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{x}{An object of class \code{VPTimeSeries}, usually a result of a call to \link[bioRad]{VPTimeSeries}}
}
\description{
print method for class \code{VPTimeSeries}
}
\keyword{internal}

