% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proj_to_wgs}
\alias{proj_to_wgs}
\title{A wrapper for \code{\link[sp:spTransform]{sp::spTransform()}}.
Converts projected coordinates to geographic (WGS84) coordinates.}
\usage{
proj_to_wgs(x, y, proj4string)
}
\arguments{
\item{x}{The x-coordinate in the projected system.}

\item{y}{The y-coordinate in the projected system.}

\item{proj4string}{An object of class 'CRS', as defined in package \code{sp}.}
}
\value{
An object of class \code{SpatialPoints}.
}
\description{
A wrapper for \code{\link[sp:spTransform]{sp::spTransform()}}.
Converts projected coordinates to geographic (WGS84) coordinates.
}
\keyword{internal}
