% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_as_corrected_ppi.R
\name{project_as_corrected_ppi}
\alias{project_as_corrected_ppi}
\title{calculate a range-bias corrected PPI}
\usage{
project_as_corrected_ppi(vol, vp, nx, ny, xlim, ylim, quantity = c("vir",
  "vid", "correction_factor", "overlap"),
  crs = "+proj=longlat +datum=WGS84", res = NA)
}
\arguments{
\item{vol}{a polar volume of class pvol}

\item{vp}{a vertical profile of class vp}

\item{nx}{number of grid pixels in the longitude dimension}

\item{ny}{number of grid pixels in the latitude dimension}

\item{xlim}{longitude range}

\item{ylim}{latitude range}

\item{quantity}{one or multiple of 'vir','vid','correction_factor', 'overlap'}
}
\value{
An object of class '\link[=summary.ppi]{ppi}'.
}
\description{
calculates a PPI that corrects for range-bias effects due to partial beam overlap with
the layer of migration (overshooting) at larger distances from the radar
}
\details{
to be written
}
\examples{
# locate example polar volume file:
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")
# load polar volume
example_pvol <- read_pvolfile(pvolfile)
# load the vertical profile for this polar volume (pre-calculated)
data(example_vp)
# calculate the range-bias corrected ppi
my_ppi <- calculate_range_corrected_ppi(example_pvol,example_vp,nx=100,ny=100,xlim=c(11,15),ylim=c(55,57))
TODO: xlim ylim aanpassen naar range
}
