% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_as_corrected_ppi.R
\name{range_correction_scan}
\alias{range_correction_scan}
\title{calculate a range-bias corrected PPI}
\usage{
range_correction_scan(scan, vp, nx, ny, xlim, ylim, param = "DBZH",
  crs = "+proj=longlat +datum=WGS84", resolution = NA)
}
\arguments{
\item{scan}{a scan (sweep) of class scan}

\item{vp}{a vertical profile of class vp}

\item{nx}{number of grid pixels in the longitude dimension}

\item{ny}{number of grid pixels in the latitude dimension}

\item{xlim}{longitude range}

\item{ylim}{latitude range}

\item{param}{reflectivity factor quantity to use, one of 'DBZH', 'DBZV', 'DBZ', 'TH', 'TV', default 'DBZH'.}
}
\value{
a SpatialPointsDataFrame
}
\description{
calculates a PPI that corrects for range-bias effects due to partial beam overlap with
the layer of migration (overshooting) at larger distances from the radar
}
\details{
to be written
}
\keyword{internal}
