% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_as_corrected_ppi.R
\name{rasterize_scan}
\alias{rasterize_scan}
\title{convert a scan in polar coordinates into a raster of class SpatialGridDataFrame in WGS84 (lat,lon) coordinates}
\usage{
rasterize_scan(scan, nx, ny, xlim, ylim, quantity = "DBZH",
  crs = "+proj=longlat +datum=WGS84", resolution = NA)
}
\description{
convert a scan in polar coordinates into a raster of class SpatialGridDataFrame in WGS84 (lat,lon) coordinates
}
\keyword{internal}
