% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{rcs}
\alias{rcs}
\alias{rcs.VP}
\alias{rcs.VPList}
\alias{rcs.VPTimeSeries}
\title{Radar cross section}
\usage{
rcs(x, ...)

\method{rcs}{VP}(x)

\method{rcs}{VPList}(x)

\method{rcs}{VPTimeSeries}(x)
}
\arguments{
\item{x}{a \code{VP}, \code{VPList} or \code{VPTimeSeries} object}
}
\value{
a radar cross section in cm^2
}
\description{
Gives the currently assumedradar cross section in cm^2.
}
\section{Methods (by class)}{
\itemize{
\item \code{VP}: radar cross section of a vertical profile

\item \code{VPList}: radar cross section of a vertical profile

\item \code{VPTimeSeries}: radar cross section of a vertical profile
}}
\examples{
# extract RCS for a single vertical profile:
rcs(VP)
}

