% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vpfiles.R
\name{read_vp}
\alias{read_vp}
\title{Read a vertical profile (\code{vp}) from file}
\usage{
read_vp(filename)
}
\arguments{
\item{filename}{A string containing the path to a vertical profile
generated by \link{calculate_vp}.}
}
\value{
An object of class \link{\code{summary.vp}}.
}
\description{
Read a vertical profile (\code{vp}) from file
}
\examples{
# locate example profile file:
vpfile <- system.file("extdata", "profile.h5", package = "bioRad")
# print the local path of the profile file:
vpfile
# load the file:
read_vp(vpfile)
}
