% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp.R
\name{readvp.list}
\alias{readvp.list}
\title{Read a list of vertical profiles from multiple files}
\usage{
readvp.list(files)
}
\arguments{
\item{files}{A string vector containing the filenames of vertical profiles in ODIM HDF5 format generated by \link{calculate_vp}}
}
\value{
an object of class \code{vplist}, which is a list \code{vp} objects
}
\description{
Read a list of vertical profiles from multiple files
}
\examples{
\dontrun{readvp(c("my/path/profile1.h5","my/path/profile2.h5", ...))}

}
