% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{readVP}
\alias{readVP}
\title{Read a vertical profile (VP) from file}
\usage{
readVP(filename)
}
\arguments{
\item{filename}{A string containing the path to a vertical profile generated by \link[bioRad]{vol2bird}}
}
\value{
an object of class \code{VP}, which is a list containing
\describe{
 \item{\code{radar}}{the radar identifier}
 \item{\code{datetime}}{the nominal time of the profile}
 \item{\code{profile}}{the profile data}
 \item{\code{attributes}}{the profile attributes}
}
}
\description{
Read a vertical profile (VP) from file
}
\examples{
# locate example profile file:
prof <- system.file("extdata", "profile.h5", package="bioRad")
# print the local path of the profile file:
prof
# load the file:
readVP(prof)

}

