% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexrad_odim.R
\name{rsl2odim}
\alias{rsl2odim}
\title{convert RSL polar volume to ODIM hdf5 format}
\usage{
rsl2odim(vol.in, vol.out, verbose = F, mount = dirname(vol.in))
}
\arguments{
\item{vol.in}{polar volume input file in RSL format}

\item{vol.out}{filename for the polar volume in ODIM hdf5 format to be generated}

\item{verbose}{logical. When TRUE, pipe Docker stdout to R console. On Windows always TRUE}

\item{mount}{character string with the mount point (a directory path) for the Docker container}
}
\value{
\code{TRUE} on success
}
\description{
convert RSL polar volume to ODIM hdf5 format
}
