% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvsd.R
\name{rvsd}
\alias{rvsd}
\alias{rvsd.vp}
\alias{rvsd.list}
\alias{rvsd.vpts}
\title{Get threshold of the VVP-retrieved radial velocity standard deviation}
\usage{
rvsd(x)

\method{rvsd}{vp}(x)

\method{rvsd}{list}(x)

\method{rvsd}{vpts}(x)
}
\arguments{
\item{x}{A \code{vp}, list of \code{vp} or \code{vpts} object.}
}
\value{
threshold for \code{sd_vvp} in m/s.
}
\description{
Gives the current threshold in VVP-retrieved radial velocity standard
deviation in m/s.
}
\section{Methods (by class)}{
\itemize{
\item \code{vp}: threshold in VVP-retrieved radial velocity standard
deviation of a vertical profile

\item \code{list}: threshold in VVP-retrieved radial velocity standard
deviation of a list of vertical profiles

\item \code{vpts}: threshold in VVP-retrieved radial velocity standard
deviation of a time series of vertical profiles
}}

\examples{
# extract threshold for a single vertical profile:
rvsd(example_vp)
}
