% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\docType{data}
\name{SCAN}
\alias{SCAN}
\title{Example object of class \code{\link[=summary.scan]{scan}}}
\format{An object of class \code{scan} of dimension 5 x 480 x 360.}
\usage{
SCAN
}
\description{
Example object of class \code{\link[=summary.scan]{scan}}
}
\keyword{datasets}

