% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SCAN}
\alias{SCAN}
\title{Example object of class \code{scan}}
\format{An object of class \code{scan} of dimension 5 x 480 x 360.}
\usage{
SCAN
}
\description{
Example of a \code{\link[=summary.scan]{scan}} object with name \code{SCAN}.
}
\examples{
# get summary of example scan:
summary(SCAN)

# SCAN was created with:
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")
pvol <- read_pvolfile(pvolfile)
SCAN <- pvol$scans[[1]]
save(SCAN, file = "data/SCAN.RData")
}
\keyword{datasets}
