% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_as_corrected_ppi.R
\name{scan_range_correction}
\alias{scan_range_correction}
\title{calculate a range-bias corrected PPI}
\usage{
scan_range_correction(scan, vp, param = "DBZH", lat = NA, lon = NA,
  antenna = NA, beam_angle = 1, k = 4/3, re = 6378, rp = 6357)
}
\arguments{
\item{scan}{a scan (sweep) of class scan}

\item{vp}{a vertical profile of class vp}

\item{param}{scan parameters to include. If \code{NA} include all scan parameters. Reducing the number
of scan parameters speeds up evaluation.}

\item{lat}{Geodetic latitude of the radar in degrees.}

\item{lon}{Geodetic longitude of the radar in degrees. If \code{NA} taken from \code{scan}.}

\item{k}{Standard refraction coefficient.}

\item{re}{Earth equatorial radius in km.}

\item{rp}{Earth polar radius in km.}
}
\value{
an object of class 'scan'
}
\description{
calculates a PPI that corrects for range-bias effects due to partial beam overlap with
the layer of migration (overshooting) at larger distances from the radar
}
\details{
to be written
}
\keyword{internal}
