% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_as_corrected_ppi.R
\name{scan_to_raster_range_correction}
\alias{scan_to_raster_range_correction}
\title{calculate a range-bias corrected PPI}
\usage{
scan_to_raster_range_correction(scan, vp, nx = 100, ny = 100,
  xlim = NA, ylim = NA, crs = "+proj=longlat +datum=WGS84",
  res = NA, param = "DBZH", lat = NA, lon = NA, antenna = NA,
  beam_angle = 1, k = 4/3, re = 6378, rp = 6357)
}
\arguments{
\item{scan}{a scan (sweep) of class scan}

\item{vp}{a vertical profile of class vp}

\item{nx}{number of raster pixels in the x (longitude) dimension}

\item{ny}{number of raster pixels in the y (latitude) dimension}

\item{xlim}{x (longitude) range}

\item{ylim}{y (latitude) range}

\item{crs}{character or object of class CRS. PROJ.4 type description of a Coordinate Reference System (map projection).
By default a WSG84 (lat,lon) projection.}

\item{res}{numeric vector of length 1 or 2 to set the resolution of the raster (see \link[raster]{res}).
If this argument is used, arguments \code{nx} and \code{ny} are ignored. Unit is identical to \code{xlim} and \code{ylim}}

\item{param}{scan parameters to include. If \code{NA} include all scan parameters. Reducing the number
of scan parameters speeds up evaluation.}

\item{lat}{Geodetic latitude of the radar in degrees.}

\item{lon}{Geodetic longitude of the radar in degrees. If \code{NA} taken from \code{scan}.}

\item{k}{Standard refraction coefficient.}

\item{re}{Earth equatorial radius in km.}

\item{rp}{Earth polar radius in km.}
}
\value{
a SpatialPointsDataFrame
}
\description{
calculates a PPI that corrects for range-bias effects due to partial beam overlap with
the layer of migration (overshooting) at larger distances from the radar
}
\details{
to be written
}
\keyword{internal}
