% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_convert.R
\name{scan_to_sf}
\alias{scan_to_sf}
\title{Convert a polar scan into a simple feature object}
\usage{
scan_to_sf(scan, lat, lon, k = 4/3, re = 6378, rp = 6357)
}
\arguments{
\item{scan}{A scan (sweep) of class \code{scan}.}
\item{lat}{Geodetic latitude of the radar in degrees. If missing, taken from \code{scan}.}
\item{lon}{Geodetic longitude of the radar in degrees. If missing, taken from \code{scan}.}
\item{k}{Numeric. Standard refraction coefficient.}
\item{re}{Numeric. Earth equatorial radius, in km.}
\item{rp}{Numeric. Earth polar radius, in km.}
}
\value{An \code{sf} object with point geometries.}
\description{
Georeferences the pixel centers of a polar scan as point geometries.
}
\details{
Beam altitude accounts for earth curvature using \code{\link[=beam_height]{beam_height()}}.
Distance over the earth's surface is calculated using \code{\link[=beam_distance]{beam_distance()}}.
}
\examples{
scan_to_sf(example_scan)
}
\seealso{
\code{\link[=scan_to_spatraster]{scan_to_spatraster()}}
}
\concept{raster functions}
