% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_convert.R
\name{scan_to_spatraster}
\alias{scan_to_spatraster}
\title{Convert a polar scan into a SpatRaster}
\usage{
scan_to_spatraster(
  scan,
  nx = 100,
  ny = 100,
  xlim,
  ylim,
  res = NA,
  param,
  raster = NA,
  lat,
  lon,
  crs = NA,
  k = 4/3,
  re = 6378,
  rp = 6357
)
}
\arguments{
\item{scan}{A scan (sweep) of class \code{scan}.}
\item{nx}{Number of raster pixels in the x dimension.}
\item{ny}{Number of raster pixels in the y dimension.}
\item{xlim}{Range in the x dimension.}
\item{ylim}{Range in the y dimension.}
\item{res}{Numeric vector of length one or two setting raster resolution.}
\item{param}{Scan parameters to include. If missing, include all scan parameters.}
\item{raster}{Optional \code{raster::RasterLayer} or \code{terra::SpatRaster} template with a CRS.}
\item{lat}{Geodetic latitude of the radar in degrees. If missing, taken from \code{scan}.}
\item{lon}{Geodetic longitude of the radar in degrees. If missing, taken from \code{scan}.}
\item{crs}{Coordinate reference system. By default, uses an azimuthal equidistant projection centered on the radar.}
\item{k}{Numeric. Standard refraction coefficient.}
\item{re}{Numeric. Earth equatorial radius, in km.}
\item{rp}{Numeric. Earth polar radius, in km.}
}
\value{A \code{terra::SpatRaster}.}
\description{
Projects a polar scan onto a regular raster grid.
}
\details{
If multiple scan pixels fall within one raster pixel, the last pixel is used.
}
\examples{
\donttest{
scan_to_spatraster(example_scan)
scan_to_spatraster(example_scan, ylim = c(55, 57), xlim = c(12, 13), res = 0.1)
}
}
\seealso{
\code{\link[=scan_to_sf]{scan_to_sf()}}
}
\concept{raster functions}
