% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{sd_vvp<-}
\alias{sd_vvp<-}
\alias{sd_vvp<-.vp}
\alias{sd_vvp<-.vplist}
\alias{sd_vvp<-.vpts}
\title{Set threshold for VVP-retrieved radial velocity standard deviation}
\usage{
sd_vvp(x) <- value

\method{sd_vvp}{vp}(x) <- value

\method{sd_vvp}{vplist}(x) <- value

\method{sd_vvp}{vpts}(x) <- value
}
\arguments{
\item{x}{a \code{vp}, \code{vplist} or \code{vpts} object}

\item{value}{the value to assign}
}
\description{
Sets the threshold in \code{sd_vvp}. Altitude layers with \code{sd_vvp} below this threshold are
assumed to have an aerial density of zero individuals. This method updates the migration densities in \code{x$data$dens}
}
\examples{
# change threshold for a single vertical profile:
sd_vvp(VP)<-2
}
