% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{sd_vvp}
\alias{sd_vvp}
\alias{sd_vvp.VP}
\alias{sd_vvp.VPList}
\alias{sd_vvp.VPTimeSeries}
\title{VVP-retrieved radial velocity standard deviation}
\usage{
sd_vvp(x, ...)

\method{rcs}{VP}(x)

\method{sd_vvp}{VPList}(x)

\method{sd_vvp}{VPTimeSeries}(x)
}
\arguments{
\item{x}{a \code{VP}, \code{VPList} or \code{VPTimeSeries} object}
}
\value{
threshold for \code{sd_vvp} in m/s.
}
\description{
Gives the current threshold in VVP-retrieved radial velocity standard deviation in m/s.
}
\section{Methods (by class)}{
\itemize{
\item \code{VP}: threshold in VVP-retrieved radial velocity standard deviation of a vertical profile

\item \code{VPList}: threshold in VVP-retrieved radial velocity standard deviation of a list of vertical profiles

\item \code{VPTimeSeries}: threshold in VVP-retrieved radial velocity standard deviation of a time series of vertical profiles
}}
\examples{
# extract threshold for a single vertical profile:
sd_vvp(VP)
}

