% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beam.R
\name{single_beam_profile}
\alias{single_beam_profile}
\title{Gaussian beam profile as a function of height relative to ground level}
\usage{
single_beam_profile(height, range, elev, k = 4/3, lat = 35,
  re = 6378, rp = 6357)
}
\arguments{
\item{height}{numeric. Height above ground level in meter. TODO: make units similar to beam_height()}

\item{range}{numeric. Range (distance from the radar antenna) in meter.}

\item{elev}{numeric. Beam elevation in degrees.}

\item{k}{Standard refraction coefficient.}

\item{lat}{Geodetic latitude in degrees.}

\item{re}{Earth equatorial radius in km.}

\item{rp}{Earth polar radius in km.}
}
\value{
numeric.
}
\description{
Normalized altitudinal pattern of radiated energy as a function of
altitude at a given distance from the radar, assuming the beam is emitted at surface level.
}
\details{
Beam profile is calculated using \link{beam_height} and \link{beam_width}.
}
\examples{
# plot the beam profile of a beam emitted at 2 degrees elevation at 35000 meter from the radar:
plot(single_beam_profile(0:3000,35000,2),0:3000,xlab="normalized radiated energy",ylab="altitude [m]",main="beam profile of 2 degree elevation beam at 35 km")
}
\keyword{internal}
