% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skip_if_no_aws.s3}
\alias{skip_if_no_aws.s3}
\title{Skip test if no aws.s3}
\usage{
skip_if_no_aws.s3()
}
\value{
Invisibly returns TRUE if aws.s3 is available, otherwise skips the test with a message "Package aws.s3 not installed".
}
\description{
Some functions require package aws.s3
This helper function allows to skip a test if aws.s3 is not available
Inspired by \url{https://testthat.r-lib.org/articles/skipping.html#helpers}.
}
\keyword{internal}
