% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{skip_if_no_docker}
\alias{skip_if_no_docker}
\title{Skip test if no docker}
\usage{
skip_if_no_docker()
}
\description{
Some functions require Docker. This helper function allows to skip a test if
the Docker container is not available, e.g. when running in CI.
Inspired by \url{https://testthat.r-lib.org/articles/skipping.html#helpers}.
}
\keyword{internal}
