% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skip_if_no_mapping}
\alias{skip_if_no_mapping}
\title{Skip test if missing dependencies for mapping}
\usage{
skip_if_no_mapping()
}
\value{
Invisibly returns TRUE if dependencies available, otherwise skips the test with a message "map() dependencies (ggspatial, prettymapr, rosm) not installed".
}
\description{
Function map depends on several spatial dependencies (ggspatial, prettymapr, rosm).
This helper function allows to skip a test if these dependencies are  not available
Inspired by \url{https://testthat.r-lib.org/articles/skipping.html#helpers}.
}
\keyword{internal}
