% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skip_if_no_mistnet}
\alias{skip_if_no_mistnet}
\title{Skip test if no mistnet}
\usage{
skip_if_no_mistnet()
}
\description{
Some functions require MistNet to be enabled in package vol2birdR.
This helper function allows to skip a test if MistNet is not available, e.g. when running in CI.
Inspired by \url{https://testthat.r-lib.org/articles/skipping.html#helpers}.
}
\keyword{internal}
