% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skip_if_no_tidyselect}
\alias{skip_if_no_tidyselect}
\title{Skip test if no tidyselect}
\usage{
skip_if_no_tidyselect()
}
\value{
Invisibly returns TRUE if tidyselect is available, otherwise skips the test with a message "Package tidyselect not installed".
}
\description{
dplyr select method require package tidyselect
This helper function allows to skip a test if tidyselect is not available
Inspired by \url{https://testthat.r-lib.org/articles/skipping.html#helpers}.
}
\keyword{internal}
