% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skip_if_no_vol2birdR}
\alias{skip_if_no_vol2birdR}
\title{Skip test if vol2birdR not installed}
\usage{
skip_if_no_vol2birdR()
}
\value{
Invisibly returns TRUE if vol2birdR is installed, otherwise skips the test with
a message "Package vol2birdR not installed".
}
\description{
Some functions require suggested package vol2birdR to be installed.
This helper function allows to skip a test if vol2birdR is not available, e.g. when running in CI.
Inspired by \url{https://testthat.r-lib.org/articles/skipping.html#helpers}.
}
\keyword{internal}
