% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spTransform2}
\alias{spTransform2}
\title{A wrapper for spTransform}
\usage{
spTransform2(...)
}
\arguments{
\item{...}{Parameters to be passed to spTransform.}
}
\value{
The result of calling spTransform with the specified parameters and
sp_evolution_status set to 2L.
}
\description{
This function calls spTransform, setting the sp_evolution_status option to 2L.
}
