% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.R
\name{[.ppi}
\alias{[.ppi}
\title{Subset a plan position indicator (\code{ppi})}
\usage{
\method{[}{ppi}(x, i)
}
\arguments{
\item{x}{A \code{ppi} object.}

\item{i}{Integer. Index/indices specifying which parameters (\code{param}) or
derived quantities to extract.}
}
\description{
Select parameters (\code{param}) or derived quantities from a plan position
indicator (\code{ppi}) by index.
}
\examples{
# Load the example scan
data(example_scan)

# Project scan as ppi
example_ppi <- project_as_ppi(example_scan)

# This ppi contains 5 parameters (VRADH DBZH ZDR RHOHV PHIDP)
example_ppi

# Subset ppi to one containing only the first parameter (VRADH)
example_ppi[1]

# Subset ppi to one containing the first three parameters (VRADH, DBZH, ZDR)
example_ppi[1:3]
}
