% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vplist.R
\name{[.vplist}
\alias{[.vplist}
\title{Subset a list of vertical profiles (\code{vplist})}
\usage{
\method{[}{vplist}(x, i)
}
\arguments{
\item{x}{object of class 'vplist'}

\item{i}{indices specifying elements to extract}
}
\description{
Extract by index from a vplist
}
