% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpts.R
\name{[.vpts}
\alias{[.vpts}
\title{Subset a time series of vertical profiles (\code{vpts})}
\usage{
\method{[}{vpts}(x, i)
}
\arguments{
\item{x}{Object of class \code{vpts}.}

\item{i}{Indices specifying elements to extract.}
}
\description{
Extract by index from a vpts
}
