% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.ppi.R
\name{summary.param}
\alias{is.param}
\alias{summary.param}
\title{Class 'param': polar scan parameter}
\usage{
\method{summary}{param}(object, ...)

is.param(x)
}
\arguments{
\item{object}{object of class 'param'}

\item{...}{additional arguments affecting the summary produced.}

\item{x}{object of class 'param'}
}
\value{
for \code{is.scan}: \code{TRUE} if its argument is of class "\code{param}"
}
\description{
Class 'param': polar scan parameter
}
\details{
Scan parameters are simple matrices, with the follwowing specific attributes:
\describe{
   \item{\code{lat}}{latitude of the radar [decimal degrees]}
   \item{\code{lon}}{longitude of the radar [decimal degrees]}
   \item{\code{height}}{height of the radar antenna [metres above sea level]}
   \item{\code{elangle}}{radar beam elevation [degrees]}
   \item{\code{param}}{string with the name of the polar scan parameter}
}
}
\examples{
is.param("this is not a polar scan parameter but a string")  #> FALSE
}

