% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{[.VPList}
\alias{[.VPList}
\alias{[.VPTimeSeries}
\alias{is.VPList}
\alias{summary.VPList}
\title{Class VPList: list of vertical profiles}
\usage{
\method{[}{VPList}(x, i)

\method{summary}{VPList}(x)

is.VPList(x)

\method{[}{VPTimeSeries}(x, i)
}
\arguments{
\item{x}{object}
}
\value{
for \code{is.VPList}: \code{TRUE} if its argument is of class "\code{VPList}"
}
\description{
Class for list of vertical profiles
}
\details{
details to be written
}
\keyword{internal}

