% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{summary.VPTimeSeries}
\alias{dim.VPTimeSeries}
\alias{is.VPTimeSeries}
\alias{summary.VPTimeSeries}
\title{Class VPTimeSeries}
\usage{
\method{summary}{VPTimeSeries}(x)

is.VPTimeSeries(x)

\method{dim}{VPTimeSeries}(x)
}
\arguments{
\item{x}{object}
}
\value{
for \code{is.VPTimeSeries}: \code{TRUE} if its argument is of class "\code{VPTimeSeries}"

for \code{dim.VPTimeSeries}: dimensions of the time series
}
\description{
Class for single-site time series of vertical profiles
}
\details{
details to be written
}

