% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpts.R
\name{summary.vpts}
\alias{summary.vpts}
\alias{print.vpts}
\alias{is.vpts}
\alias{dim.vpts}
\title{Inspect a time series of vertical profiles (\code{vpts})}
\usage{
\method{summary}{vpts}(object, ...)

\method{print}{vpts}(x, digits = max(3L, getOption("digits") - 3L), references = TRUE, ...)

is.vpts(x)

\method{dim}{vpts}(x)
}
\arguments{
\item{object}{A \code{vpts} object.}

\item{...}{Additional arguments affecting the summary produced.}

\item{x}{A \code{vpts} object.}

\item{digits}{minimal number of \emph{significant} digits, see
    \code{\link[base]{print.default}}.}

\item{references}{Logical indicating whether to print the references for the vpts
object. If \code{TRUE} and the object contains references in
\code{x$attributes$references}, these will be printed in author-year format.
Default is \code{TRUE}.}
}
\value{
For \code{\link[=summary.vpts]{summary.vpts()}}: prints the summary of the\code{vpts} object.

For \code{\link[=print.vpts]{print.vpts()}}: prints the summary of the \code{vpts} object.

For \code{\link[=is.vpts]{is.vpts()}}: \code{TRUE} for an object of class \code{vpts}, otherwise
\code{FALSE}.

For \code{\link[=dim.vpts]{dim.vpts()}}: number of datetimes, heights and quantities in a
time series of vertical profiles (\code{vpts}).
}
\description{
R base functions for inspecting a time series of vertical profiles (\code{vpts})
object.
}
\details{
A time series of vertical profiles contains time-ordered vertical profiles
(\verb{vp)} of a single radar. This time series can be \strong{regular} (\code{vp} are
equally spaced in time) or \strong{irregular} (time steps between \code{vp} are of
unequal length), indicated in the field \code{regular}. Irregular time series can
be projected onto a regular time grid with \code{\link[=regularize_vpts]{regularize_vpts()}}. A time series
of vertical profile (\code{vp}) object is a list containing:
\itemize{
\item \code{radar}: Radar identifier.
\item \code{datetime}: Nominal times of the profiles (named \code{dates} in bioRad <
0.4.0) in UTC.
\item \code{height}: Lowest height of the height bins in the profiles in m above sea
level.
\item \code{daterange}: Minimum and maximum nominal time of the profiles in UTC.
\item \code{timesteps}: Time differences between the profiles. Element \code{i} gives the
difference between profile \code{i} and \code{i+1}.
\item \code{data}: A list of quantities, each containing a \code{datetime} by \code{height}
matrix with the values. Use \code{\link[=get_quantity]{get_quantity()}} to access these and see
\code{\link[=summary.vp]{summary.vp()}} for a description of available quantities.
\item \code{attributes}: List of the vertical profile's \code{what}, \code{where}, and \code{how}
attributes, copied from the first profile.
\item \code{regular}: Logical indicating whether the time series is regular or not.
}
}
\section{Conventions}{

\itemize{
\item \code{NA}: Maps to \code{nodata} in the ODIM convention: value to denote areas void
of data (never radiated).
\item \code{NaN}: Maps to \code{undetect} in the ODIM convention: denote areas below the
measurement detection threshold (radiated but nothing detected). The value is
also used when there are too few datapoints to calculate a quantity.
\item \code{0}: Maps to \code{0} in the ODIM convention: denote areas where the quantity
has a measured value of zero (radiated and value zero detected or inferred).
}
}

\examples{
# Check if an object is of class vpts
is.vpts(example_vpts)

# Get summary info
example_vpts # Same as summary(example_vpts) or print(example_vpts)

# Get dimensions
dim(example_vpts)

# To suppress references printing
print(example_vpts, references = FALSE)
}
\seealso{
Other vpts functions:
\code{\link[=[.vpts]{[.vpts()}},
\code{\link{example_vpts}},
\code{\link[=plot.vpts]{plot.vpts()}}
}
\concept{vpts functions}
