% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunrise_sunset.R
\name{suntime}
\alias{suntime}
\title{Helper function to caclulate sunrise or sunset}
\usage{
suntime(lon, lat, date, elev = -0.268, rise = TRUE)
}
\arguments{
\item{lon}{Longitude in decimal degrees.}

\item{lat}{Latitude in decimal degrees.}

\item{date}{Date inhereting from class \code{POSIXt} or a string
interpretable by \link[base]{as.Date}.}

\item{elev}{Sun elevation in degrees.}

\item{rise}{Whether to output for rising (\code{TRUE}) or setting
(\code{FALSE}) sun.}
}
\description{
Use the \code{\link{sunrise}} or \code{\link{sunset}} functions. The current
approach with rise = TRUE is kept for backward compatibility
}
