% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp_repo.R
\name{url_existence}
\alias{url_existence}
\title{Check is a specific URL is existing and providing a return message}
\usage{
url_existence(url)
}
\arguments{
\item{url}{char weblink to a potentially existing webpage}
}
\value{
z with a lenght > 1 if the URL is existing and downloading would be
possible
}
\description{
Check is a specific URL is existing and providing a return message
}
\keyword{internal}
