% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vp.R
\name{url_existencs}
\alias{url_existencs}
\title{Check is a specific URL is existing and providing a return message}
\usage{
url_existencs(url)
}
\arguments{
\item{url}{char weblink to a potentially existing webpage}
}
\value{
z with a lenght > 1 if the URL is existing and downloading would be
possible
}
\description{
Check is a specific URL is existing and providing a return message
}

