% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_vpts.R
\name{validate_vpts}
\alias{validate_vpts}
\title{Validate dataframe against VPTS schema}
\usage{
validate_vpts(df)
}
\arguments{
\item{df}{The dataframe to validate.}
}
\value{
Invisibly returns a list with validation results including messages for any issues found.
}
\description{
Validate dataframe against VPTS schema
}
\examples{
my_vpts <- as.data.frame(example_vpts)
validate_vpts(my_vpts)
}
\keyword{internal}
