% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcp.R
\name{vcp}
\alias{vcp}
\title{Extract a volume coverage pattern table}
\usage{
vcp(
  x,
  add_params = T,
  select = c("how.lowprf", "how.midprf", "how.highprf", "where.elangle", "where.nbins",
    "where.nrays", "where.rscale", "how.NI"),
  ...
)
}
\arguments{
\item{x}{Either a pvol or scan for which the VCP should be created.}

\item{add_params}{A logical wether a column with parameters be added.}

\item{select}{A charcter vector which the column names that should be returned when NULL all attributes are returned}

\item{...}{Currently not used

This function tabulates the attributes of one scan or all scans of a pvol. Attributes that have a length longer then one are presented as a list column.
By default the function returns a limited set of column to keep the output clear.
It is important to note thad attributes of the polar volume can contain additional information on processing that is not included in the resulting table.}
}
\description{
Extract a volume coverage pattern table
}
\examples{
data(example_scan)
vcp(example_scan)

pvolfilepath <- system.file("extdata", "volume.h5", package = "bioRad")
example_pvol <- read_pvolfile(pvolfilepaht)
vcp(example_pvol)
}
