% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{vol2bird_version}
\alias{vol2bird_version}
\title{Check version of the vol2bird algorithm used by bioRad}
\usage{
vol2bird_version(vol2bird_local_install)
}
\arguments{
\item{vol2bird_local_install}{(optional) String with path to local
vol2bird installation, see \link{calculate_vp} for details.}
}
\value{
an object of class \link{numeric_version}
}
\description{
Checks that \href{https://www.docker.com/}{Docker} daemon is running
correctly on the local system and returns the version of the
installed vol2bird algorithm in the Docker container.
}
\details{
when argument \code{vol2bird_local_install} is specified with
a path to a local executable of vol2bird, the function will return
the version of this local installation.
}
