% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\docType{data}
\name{VP}
\alias{VP}
\title{Example object of class \code{\link[=summary.vp]{vp}} as generated by \code{\link{vol2bird}} or \code{\link{readvp}}}
\format{An object of class \code{vp} with 20 rows and 16 columns.}
\usage{
VP
}
\description{
Example object of class \code{\link[=summary.vp]{vp}} as generated by \code{\link{vol2bird}} or \code{\link{readvp}}
}
\keyword{datasets}

