% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{VP}
\alias{VP}
\title{Example object of class \code{vp}}
\format{An object of class \code{vp} with 20 rows and 16 columns.}
\usage{
VP
}
\description{
Example of a \code{\link[=summary.vp]{vp}} object with name \code{VP}. Can
be created with \code{\link{calculate_vp()}} or read from file with
\code{\link{read_vpfiles()}}.
}
\examples{
# get summary of example vp:
summary(VP)

# VP was created with:
VP <- read_vpfiles("/inst/extdata/profile.h5")
save(VP, file = "data/VP.RData")
}
\keyword{datasets}
