% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{VPFileQ}
\alias{VPFileQ}
\title{Check whether file is a vertical profile}
\usage{
VPFileQ(filename)
}
\arguments{
\item{filename}{A string containing a filename}
}
\value{
TRUE when \code{filename} is a vertical profile, otherwise FALSE
}
\description{
Checker whether a file is a vertical profile that can be read with package \pkg{bioRad}
}
\examples{
VPFileQ("this/is/not/a/valid/VP/file.h5")   #> FALSE

}

