% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{VPTS}
\alias{VPTS}
\title{Example object of class \code{vpts}}
\format{An object of class \code{vpts} of dimension 25 x 6135 x 15.}
\usage{
VPTS
}
\description{
Example object of class \code{vpts}, a time series of vertical profiles
}
\keyword{datasets}
