% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\docType{data}
\name{VPTS}
\alias{VPTS}
\title{Example object of class \code{VPTimeSeries}}
\format{An object of class \code{VPTimeSeries} of dimension 25 x 6135 x 15.}
\usage{
VPTS
}
\description{
Example object of class \code{VPTimeSeries}
}
\keyword{datasets}

