% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{VPTS}
\alias{VPTS}
\title{Example object of class \code{vpts}}
\format{An object of class \code{vpts} of dimension 25 x 6135 x 15.}
\usage{
VPTS
}
\description{
Example of a \code{\link[=summary.vpts]{vpts}} object (a time series of
vertical profiles) with name \code{SCAN}.
}
\examples{
# get summary of example vpts:
summary(VPTS)

# VPTS was created with:
vptsfile <- system.file("extdata", "VPtable.txt", package = "bioRad")
VPTS <- read_vpts(vptsfile, radar = "KBGM", wavelength = "S")
rcs(VPTS) <- 11
VPTS$attributes$where$lat <- 42.2
VPTS$attributes$where$lon <- -75.98
save(VPTS, file = "data/VPTS.RData", compress = "xz")
}
\keyword{datasets}
