% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vpts_schema}
\alias{vpts_schema}
\title{ENRAM-defined VPTS schema}
\format{
An object of class \code{list} of length 3.
}
\usage{
vpts_schema
}
\value{
A list object containing the data model for the standard VPTS exchange format.
}
\description{
A dataset containing the schema definitions for VPTS data validation.
}
\examples{
# Reload vpts_schema from package (e.g. in case it was altered)
data(vpts_schema)
}
\keyword{data}
