% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wgs_to_proj}
\alias{wgs_to_proj}
\title{A wrapper for \code{\link[sp:spTransform]{sp::spTransform()}}.
Converts geographic (WGS84) coordinates to a specified projection}
\usage{
wgs_to_proj(lon, lat, proj4string)
}
\arguments{
\item{lon}{Longitude}

\item{lat}{Latitude}

\item{proj4string}{An object of class 'CRS', as defined in package \code{sp}.}
}
\value{
An object of class \code{SpatialPoints}.
}
\description{
A wrapper for \code{\link[sp:spTransform]{sp::spTransform()}}.
Converts geographic (WGS84) coordinates to a specified projection
}
\keyword{internal}
