% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_as_corrected_ppi.R
\name{add_expected_eta_to_scan}
\alias{add_expected_eta_to_scan}
\title{calculate a range-bias corrected PPI}
\usage{
add_expected_eta_to_scan(scan, vp, param = "DBZH", lat = NA,
  lon = NA, antenna = NA, beam_angle = 1, k = 4/3, re = 6378,
  rp = 6357)
}
\arguments{
\item{scan}{a scan (sweep) of class scan}

\item{vp}{a vertical profile of class vp}

\item{param}{reflectivity factor scan parameter on which to base range bias corrections.
Typically the same parameter from which animal densities are estimated for object \code{vp}.
One of 'DBZH','DBZV','DBZ','TH','TV'.}

\item{lat}{Geodetic latitude of the radar in degrees. If \code{NA} taken from \code{pvol}.}

\item{lon}{Geodetic latitude of the radar in degrees. If \code{NA} taken from \code{pvol}.}

\item{beam_angle}{numeric. Beam opening angle in degrees, typically the
the angle between the half-power (-3 dB) points of the main lobe}

\item{k}{Standard refraction coefficient.}

\item{re}{Earth equatorial radius in km.}

\item{rp}{Earth polar radius in km.}
}
\value{
an object of class 'scan'
}
\description{
calculates a PPI that corrects for range-bias effects due to partial beam overlap with
the layer of biological echoes (overshooting) at larger distances from the radar
}
\details{
to be written
}
\keyword{internal}
