% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_param.R
\name{add_param}
\alias{add_param}
\alias{add_param.scan}
\alias{add_param.pvol}
\title{Add scan parameter from a georeferenced raster}
\usage{
add_param(x, raster, param)

\method{add_param}{scan}(x, raster, param)

\method{add_param}{pvol}(x, raster, param)
}
\arguments{
\item{x}{A \code{pvol} or \code{scan} object.}

\item{raster}{An object of class \code{terra::SpatRaster} or \code{raster::RasterLayer}.}

\item{param}{The name of the added parameter.}
}
\value{
The object \code{x} with an added parameter, extracting data from the raster
specified by \code{raster}.
}
\description{
Add scan parameter from a georeferenced raster
}
