% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beam.R
\name{beam_profile}
\alias{beam_profile}
\title{Calculate vertical radiation profile}
\usage{
beam_profile(height, range, elev, beam_angle = 1, k = 4/3, lat = 35,
  re = 6378, rp = 6357)
}
\arguments{
\item{height}{numeric. Height(s) above ground level in meter. TODO: make units similar to beam_height()}

\item{range}{Distance from the radar (over ground level) for which to calculate the altitudinal beam profile in m.}

\item{elev}{numeric vector of radar beam elevation(s) in degrees.}

\item{beam_angle}{numeric. Beam opening angle in degrees, typically the
the angle between the half-power (-3 dB) points of the main lobe}

\item{k}{Standard refraction coefficient.}

\item{lat}{Geodetic latitude in degrees.}

\item{re}{Earth equatorial radius in km.}

\item{rp}{Earth polar radius in km.}
}
\value{
numeric vector. Normalized radiated energy at each of the specified heights.
}
\description{
Calculate for a set of beam elevations elev
the altitudinal normalized distribution of radiated energy by those beams.
}
\details{
Beam profile is calculated using \link{beam_height} and \link{beam_width}.
Returns a beam profile as a function of height relative to ground level.

Returns the normalized altitudinal pattern of radiated energy as a function of
altitude at a given distance from the radar, assuming the beams are emitted at surface level.
}
\examples{
plot(beam_profile(0:3000,35000,c(1,2)),0:3000,xlab="normalized radiated energy",ylab="altitude [m]",main="beam elevations: 1,2 degrees; range: 35 km")
}
