% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beam.R
\name{beam_width}
\alias{beam_width}
\title{Calculate radar beam width}
\usage{
beam_width(range, beam_angle = 1)
}
\arguments{
\item{range}{numeric. Range (distance from the radar antenna) in km.}

\item{beam_angle}{numeric. Beam opening angle in degrees, typically the
the angle between the half-power (-3 dB) points of the main lobe}
}
\value{
numeric. Beam width in m.
}
\description{
Calculates the width of a radar beam as a function of range and beam angle.
}
