% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad-deprecated.R
\name{bioRad-deprecated}
\alias{bioRad-deprecated}
\alias{beamheight-deprecated}
\alias{beamheight}
\alias{beamwidth-deprecated}
\alias{beamwidth}
\alias{bind-deprecated}
\alias{bind}
\alias{vol2bird-deprecated}
\alias{vol2bird}
\alias{night-deprecated}
\alias{night}
\alias{day-deprecated}
\alias{day}
\alias{composite-deprecated}
\alias{composite}
\alias{dbz2eta-deprecated}
\alias{dbz2eta}
\alias{eta2dbz-deprecated}
\alias{eta2dbz}
\title{Deprecated functions in package \pkg{bioRad}.}
\usage{
beamheight(range, elev, k = 4/3, lat = 35, re = 6378, rp = 6357)

beamwidth(range, angle = 1)

bind(x, ...)

vol2bird(...)

night(x, ..., elev = -0.268)

day(x, elev = -0.268)

composite(...)

dbz2eta(...)

eta2dbz(...)
}
\description{
The functions listed below are deprecated or renamed and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned.
}
\section{\code{beamheight}}{

For \code{beamheight}, use \code{\link{beam_height}}.
}

\section{\code{beamwidth}}{

For \code{beamwidth}, use \code{\link{beam_width}}.
}

\section{\code{bind}}{

For \code{bind}, use \code{\link{bind_into_vpts}}.
}

\section{\code{vol2bird}}{

For \code{vol2bird}, use \code{\link{calculate_vp}}.
}

\section{\code{night}}{

For \code{night}, use \code{\link{check_night}}.


For \code{composite}, use \code{\link{composite_ppi}}.


For \code{dbz2eta}, use \code{\link{dbz_to_eta}}.


For \code{eta2dbz}, use \code{\link{eta_to_dbz}}.
}

\section{\code{day}}{

For \code{day}, use \code{\link{check_night}}.
}

\keyword{internal}
