% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad-deprecated.R
\name{bioRad-deprecated}
\alias{bioRad-deprecated}
\alias{beamheight-deprecated}
\alias{beamheight}
\alias{beamwidth-deprecated}
\alias{beamwidth}
\alias{bind-deprecated}
\alias{bind}
\title{Deprecated functions in package \pkg{bioRad}.}
\usage{
beamheight(range, elev, k = 4/3, lat = 35, re = 6378, rp = 6357)

beamwidth(range, angle = 1)

bind(x, ...)
}
\description{
The functions listed below are deprecated or renamed and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned.
}
\section{\code{beamheight}}{

For \code{beamheight}, use \code{\link{beam_height}}.
}

\section{\code{beamwidth}}{

For \code{beamwidth}, use \code{\link{beam_width}}.
}

\section{\code{bind}}{

For \code{bind}, use \code{\link{bind_into_vpts}}.
}

\keyword{internal}
