% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad-deprecated.R
\name{bioRad-deprecated}
\alias{bioRad-deprecated}
\title{Deprecated functions in package \pkg{bioRad}.}
\description{
The functions listed below are deprecated or renamed and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned.
}
\keyword{internal}
