% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad-deprecated.R
\name{bioRad-deprecated}
\alias{bioRad-deprecated}
\alias{basemap}
\alias{beamheight}
\alias{beamwidth}
\alias{bind}
\alias{checkDocker}
\alias{composite}
\alias{cmt}
\alias{day}
\alias{dbz2eta}
\alias{download_vp}
\alias{elangle}
\alias{eta2dbz}
\alias{fetch}
\alias{getscan}
\alias{h5ODIMobject}
\alias{is.vplist}
\alias{mt}
\alias{mtr}
\alias{night}
\alias{ppi}
\alias{read.pvol}
\alias{readvp}
\alias{readvp.list}
\alias{readvp.table}
\alias{regularize}
\alias{retrieve_vp_paths}
\alias{rsl2odim}
\alias{sd_vvp<-}
\alias{suntime}
\alias{updateDocker}
\alias{vintegrate}
\alias{vol2bird}
\alias{vpts}
\alias{VP-deprecated}
\alias{VPTS-deprecated}
\alias{SCAN-deprecated}
\title{Deprecated functions and data in package \pkg{bioRad}}
\usage{
basemap(...)

beamheight(range, elev, k = 4/3, lat = 35, re = 6378, rp = 6357)

beamwidth(range, angle = 1)

bind(x, ...)

checkDocker(...)

composite(...)

cmt(x, alt.min = 0, alt.max = Inf, alpha = NA, interval.max = Inf)

day(x, elev = -0.268)

dbz2eta(...)

download_vp(...)

elangle(...)

eta2dbz(...)

fetch(...)

getscan(...)

h5ODIMobject(...)

is.vplist(...)

mt(x, alt.min = 0, alt.max = Inf, alpha = NA, interval.max = Inf)

mtr(x, alt.min = 0, alt.max = Inf, alpha = NA)

night(x, ..., elev = -0.268)

ppi(...)

read.pvol(...)

readvp(...)

readvp.list(...)

readvp.table(...)

regularize(...)

retrieve_vp_paths(...)

rsl2odim(...)

sd_vvp(x) <- value

suntime(..., rise = TRUE)

updateDocker(...)

vintegrate(...)

vol2bird(...)

vpts(...)
}
\description{
The functions and data listed below are deprecated or renamed and will be defunct in
the near future. When possible, alternative functions with similar
functionality are mentioned.
}
\section{basemap}{

Use \code{\link{download_basemap}} instead.
}

\section{beamheight}{

Use \code{\link{beam_height}} instead.
}

\section{beamwidth}{

Use \code{\link{beam_width}} instead.
}

\section{bind}{

Use \code{\link{bind_into_vpts}} instead.
}

\section{checkDocker}{

Use \code{\link{check_docker}} instead.
}

\section{composite}{

Use \code{\link{composite_ppi}} instead.
}

\section{cmt}{

\code{\link{cmt}} is deprecated as a separate function. Cumulative migration
traffic is now included in the output of \code{\link{integrate_profile}} as
column \code{mt}, which can be summed to get the cumulative migration
traffic.
}

\section{day}{

Use \code{\link{check_night}} instead.
}

\section{dbz2eta}{

Use \code{\link{dbz_to_eta}} instead.
}

\section{download_vp}{

Use \code{\link{download_vpfiles}} instead.
}

\section{elangle}{

Use \code{\link{get_angles}} instead.
}

\section{eta2dbz}{

Use \code{\link{eta_to_dbz}} instead.
}

\section{fetch}{

Use \code{\link{get_quantity}} instead.
}

\section{getscan}{

Use \code{\link{get_scan}} instead.
}

\section{h5ODIMobject}{

Use \code{\link{get_odim_object_type}} instead.
}

\section{is.vplist}{

The \code{vplist} object is deprecated. Use a regular list of \code{vp}
objects (\code{c(vp, vp))} instead.
}

\section{mt}{

\code{\link{mt}} is deprecated as a separate function. Migration traffic
is now included in the output of \code{\link{integrate_profile}} as
column \code{mt}.
}

\section{mtr}{

\code{\link{mtr}} is deprecated as a separate function. Migration traffic
rate is now included in the output of \code{\link{integrate_profile}} as
column \code{mtr}.
}

\section{night}{

Use \code{\link{check_night}} instead.
}

\section{ppi}{

Use \code{\link{project_as_ppi}} instead.
}

\section{read.pvol}{

Use \code{\link{read_pvolfile}} instead.
}

\section{readvp}{

Use \code{\link{read_vpfiles}} instead.
}

\section{readvp.list}{

Use \code{\link{read_vpfiles}} instead.
}

\section{readvp.table}{

Use \code{\link{read_vpts}} instead.
}

\section{regularize}{

Use \code{\link{regularize_vpts}} instead.
}

\section{retrieve_vp_paths}{

Use \code{\link{select_vpfiles}} instead.
}

\section{rsl2odim}{

Use \code{\link{nexrad_to_odim}} instead.
}

\section{sd_vvp<-}{

Use \code{\link{rvsd<-}} instead.
}

\section{suntime}{

Use \code{\link{sunrise}} or \code{\link{sunset}} instead.
}

\section{updateDocker}{

Use \code{\link{update_docker}} instead.
}

\section{vintegrate}{

Use \code{\link{integrate_profile}} instead.
}

\section{vol2bird}{

Use \code{\link{calculate_vp}} instead.
}

\section{vpts}{

Use \code{\link{bind_into_vpts}} instead.
}

\section{\code{VP}}{

Use \code{\link{example_vp}} instead.
}

\section{\code{VPTS}}{

Use \code{\link{example_vpts}} instead.
}

\section{\code{SCAN}}{

Use \code{\link{example_scan}} instead.
}

\keyword{internal}
