% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad-deprecated.R
\name{bioRad-deprecated}
\alias{bioRad-deprecated}
\alias{beamheight-deprecated}
\alias{beamheight}
\alias{beamwidth-deprecated}
\alias{beamwidth}
\alias{bind-deprecated}
\alias{bind}
\alias{vol2bird-deprecated}
\alias{vol2bird}
\alias{night-deprecated}
\alias{night}
\alias{day-deprecated}
\alias{day}
\alias{composite-deprecated}
\alias{composite}
\alias{dbz2eta-deprecated}
\alias{dbz2eta}
\alias{eta2dbz-deprecated}
\alias{eta2dbz}
\alias{checkDocker-deprecated}
\alias{checkDocker}
\alias{updateDocker-deprecated}
\alias{updateDocker}
\alias{basemap-deprecated}
\alias{basemap}
\alias{download_vp-deprecated}
\alias{download_vp}
\alias{elangle-deprecated}
\alias{elangle}
\alias{fetch-deprecated}
\alias{fetch}
\alias{getscan-deprecated}
\alias{getscan}
\alias{vintegrate-deprecated}
\alias{vintegrate}
\title{Deprecated functions in package \pkg{bioRad}.}
\usage{
beamheight(range, elev, k = 4/3, lat = 35, re = 6378, rp = 6357)

beamwidth(range, angle = 1)

bind(x, ...)

vol2bird(...)

night(x, ..., elev = -0.268)

day(x, elev = -0.268)

composite(...)

dbz2eta(...)

eta2dbz(...)

checkDocker(...)

updateDocker(...)

basemap(...)

download_vp(...)

elangle(...)

fetch(...)

getscan(...)

vintegrate(...)
}
\description{
The functions listed below are deprecated or renamed and will be defunct in
the near future. When possible, alternative functions with similar
functionality are mentioned.
}
\section{\code{beamheight}}{

Use \code{\link{beam_height}} instead.
}

\section{\code{beamwidth}}{

Use \code{\link{beam_width}} instead.
}

\section{\code{bind}}{

Use \code{\link{bind_into_vpts}} instead.
}

\section{\code{vol2bird}}{

Use \code{\link{calculate_vp}} instead.
}

\section{\code{night}}{

Use \code{\link{check_night}} instead.
}

\section{\code{day}}{

Use \code{\link{check_night}} instead.
}

\section{\code{composite}}{

Use \code{\link{composite_ppi}} instead.
}

\section{\code{dbz2eta}}{

Use \code{\link{dbz_to_eta}} instead.
}

\section{\code{eta2dbz}}{

Use \code{\link{eta_to_dbz}} instead.
}

\section{\code{checkDocker}}{

Use \code{\link{check_docker}} instead.
}

\section{\code{updateDocker}}{

Use \code{\link{update_docker}} instead.
}

\section{\code{basemap}}{

For \code{basemap}, use \code{\link{download_basemap}}.
}

\section{\code{download_vp}}{

For \code{download_vp}, use \code{\link{download_vpfiles}}.
}

\section{\code{elangle}}{

For \code{elangle}, use \code{\link{get_angles}}.
}

\section{\code{fetch}}{

For \code{fetch}, use \code{\link{get_quantity}}.
}

\section{\code{getscan}}{

For \code{getscan}, use \code{\link{get_scan}}.
}

\section{\code{vintegrate}}{

For \code{vintegrate}, use \code{\link{integrate_profile}}.
}

\keyword{internal}
