% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\docType{package}
\name{bioRad-package}
\alias{bioRad}
\alias{bioRad-package}
\title{Analyze and visualize weather radar data}
\description{
Analyze and visualize weather radar data
}
\references{
Dokter 2010 Interface
}

