% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\docType{package}
\name{bioRad-package}
\alias{bioRad}
\alias{bioRad-package}
\title{bioRad: Biological Analysis and Visualization of Weather Radar Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extract, visualize and summarize aerial movements of birds and insects from weather radar data. See \doi{10.1111/ecog.04028} for a software paper describing package and methodologies.
}
\details{
To get started, see:
\itemize{
\item Dokter et al. (2019) \doi{10.1111/ecog.04028}: a paper describing the
package.
\item \href{https://adriaandokter.com/bioRad/articles/bioRad.html}{bioRad vignette}: an
introduction to bioRad's main functionalities.
\item \href{https://adriaandokter.com/bioRad/reference/index.html}{Function reference}: an
overview of all bioRad functions.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adokter/bioRad/}
  \item \url{https://adriaandokter.com/bioRad/}
  \item Report bugs at \url{https://github.com/adokter/bioRad/issues}
}

}
\author{
\strong{Maintainer}: Adriaan M. Dokter \email{amd427@cornell.edu} (\href{https://orcid.org/0000-0001-6573-066X}{ORCID})

Authors:
\itemize{
  \item Peter Desmet \email{peter.desmet@inbo.be} (\href{https://orcid.org/0000-0002-8442-8025}{ORCID})
  \item Stijn Van Hoey \email{stijnvanhoey@gmail.com} (\href{https://orcid.org/0000-0001-6413-3185}{ORCID})
}

Other contributors:
\itemize{
  \item Bart Hoekstra \email{b.d.hoekstra@uva.nl} (\href{https://orcid.org/0000-0002-7085-3805}{ORCID}) [contributor]
  \item Bart Kranstauber \email{b.kranstauber@uva.nl} (\href{https://orcid.org/0000-0001-8303-780X}{ORCID}) [contributor]
  \item Hidde Leijnse \email{hidde.leijnse@knmi.nl} (\href{https://orcid.org/0000-0001-7835-4480}{ORCID}) [contributor]
  \item Cecilia Nilsson \email{cecilia.nilsson709@gmail.com} (\href{https://orcid.org/0000-0001-8957-4411}{ORCID}) [contributor]
  \item Nicolas Noé  \email{nicolas@niconoe.eu} (\href{https://orcid.org/0000-0002-9503-4750}{ORCID}) [contributor]
  \item Raphaël Nussbaumer \email{rafnuss@gmail.com} (\href{https://orcid.org/0000-0002-8185-1020}{ORCID}) [contributor]
  \item Jurriaan Spaaks \email{j.spaaks@esciencecenter.nl} [contributor]
  \item Lourens Veen \email{l.veen@esciencecenter.nl} [contributor]
  \item Liesbeth Verlinden \email{l.verlinden@uva.nl} (\href{https://orcid.org/0000-0003-1744-9325}{ORCID}) [contributor]
}

}
\keyword{internal}
