% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vplist.R
\name{c.vp}
\alias{c.vp}
\title{Concatenate vertical profiles (\code{vp}) into a list of vertical profiles (\code{vplist})}
\usage{
\method{c}{vp}(...)
}
\arguments{
\item{...}{objects of class \code{vp}}
}
\value{
an object of class \code{vplist}, see \link{readvp.list} for details
}
\description{
Concatenate vertical profiles (\code{vp}) into a list of vertical profiles (\code{vplist})
}
