% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp.R
\name{c.vp}
\alias{c.vp}
\title{Concatenate vertical profiles (\code{vp}) into a list of vertical profiles}
\usage{
\method{c}{vp}(...)
}
\arguments{
\item{...}{objects of class \code{vp}}
}
\value{
an object of class \code{list}
}
\description{
Concatenate vertical profiles (\code{vp}) into a list of vertical profiles
}
