% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{c.VP}
\alias{c.VP}
\title{concatenate \code{VP} objects into a \code{VPList} object}
\usage{
\method{c}{VP}(...)
}
\arguments{
\item{...}{objects of class \code{VP}}
}
\value{
an object of class \code{VPList}, see \link[bioRad]{readVPList} for details
}
\description{
concatenate \code{VP} objects into a \code{VPList} object
}

