% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{c.vp}
\alias{c.vp}
\title{concatenate \code{vp} objects into a \code{vplist} object}
\usage{
\method{c}{vp}(...)
}
\arguments{
\item{...}{objects of class \code{vp}}
}
\value{
an object of class \code{vplist}, see \link{readvp.list} for details
}
\description{
concatenate \code{vp} objects into a \code{vplist} object
}
