% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp.R
\name{c.vp}
\alias{c.vp}
\title{Concatenate vertical profiles (\code{vp}) into a list of vertical profiles}
\usage{
\method{c}{vp}(...)
}
\arguments{
\item{...}{\code{vp} objects.}
}
\value{
A list of \code{vp} objects.
}
\description{
Concatenates vertical profiles (\code{vp}) into a list of vertical profiles
(\code{c(vp, vp, vp)}) and warns if they are not from a single radar.
}
\seealso{
\code{\link[=bind_into_vpts]{bind_into_vpts()}}
}
