% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_param.R
\name{calculate_param}
\alias{calculate_param}
\alias{calculate_param.pvol}
\alias{calculate_param.ppi}
\alias{calculate_param.scan}
\title{Calculate a new scan parameter}
\usage{
calculate_param(x, ...)

\method{calculate_param}{pvol}(x, ...)

\method{calculate_param}{ppi}(x, ...)

\method{calculate_param}{scan}(x, ...)
}
\arguments{
\item{x}{an object of class \code{pvol} or class \code{scan}}

\item{...}{an expression defining the new scan parameter in terms of existing scan parameters}
}
\value{
an object of the same class as \code{x}, either class \code{pvol} or class \code{scan}
}
\description{
Calculates a new scan parameter from a combination of existing scan parameters. Useful
for calculating quantities that are defined in terms of other basic radar moments, like
linear reflectivity eta, depolarization ratio (Kilambi et al. 2018), or for applying
clutter corrections (CCORH) to uncorrected reflectivity moments (TH), as in TH+CCORH
}
\section{Methods (by class)}{
\itemize{
\item \code{calculate_param(pvol)}: Calculate a new scan parameter for all scans in a polar volume.

\item \code{calculate_param(ppi)}: Calculate a new parameter for a PPI.

\item \code{calculate_param(scan)}: Calculate a new scan parameter for a scan

}}
\examples{
# locate example volume file:
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")

# load the file:
example_pvol <- read_pvolfile(pvolfile)
data(example_scan)

# calculate linear reflectivity ETA from reflectivity factor DBZH:
radar_wavelength <- example_pvol$attributes$how$wavelength
# example_pvol <- calculate_param(example_pvol,ETA=dbz_to_eta(DBZH,radar_wavelength))

# add depolarization ratio (DR) as a scan parameter (see Kilambi 2018):
example_pvol <- calculate_param(example_pvol, DR = 10 * log10((ZDR + 1 - 2 * ZDR^0.5 * RHOHV) /
  (ZDR + 1 + 2 * ZDR^0.5 * RHOHV)))

# calculate_param operates on both pvol and scan objects:
calculate_param(example_scan, DR = 10 * log10((ZDR + 1 - 2 * ZDR^0.5 * RHOHV) /
  (ZDR + 1 + 2 * ZDR^0.5 * RHOHV)))

# it also works for ppis
ppi <- project_as_ppi(example_scan)
calculate_param(ppi, exp(DBZH))

}
\references{
\itemize{
\item Kilambi, A., Fabry, F., and Meunier, V., 2018. A simple and effective
method for separating meteorological from nonmeteorological targets using
dual-polarization data. Journal of Atmospheric and Oceanic Technology, 35,
1415–1424. \doi{10.1175/JTECH-D-17-0175.1}
}
}
