% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vpfiles.R
\name{check_date_format}
\alias{check_date_format}
\title{Check if character date is in specific format}
\usage{
check_date_format(date, format)
}
\arguments{
\item{date}{character. Character representation of a date, e.g.
\code{"2018-12-13"}.}

\item{format}{character. strptime format the date should have, e.g.
\code{"\\\%Y-\\\%m-\\\%d"}}
}
\value{
NULL. Will stop and show error message if date does not have correct
date format.
}
\description{
Check if character date is in specific format
}
\keyword{internal}
