% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{check_docker}
\alias{check_docker}
\title{Check if Docker is running}
\usage{
check_docker(verbose = TRUE)
}
\arguments{
\item{verbose}{logical. When TRUE messages are printed to
R console.}
}
\value{
0 upon success, otherwise an error code: 1 if Docker vol2bird image not available,
2 if Docker daemon not running, 3 if Docker daemon not found.
}
\description{
Checks that \href{https://www.docker.com/}{Docker} daemon is running
correctly on the local system, and that vol2bird Docker image is available.
}
\examples{
# check if Docker is running and vol2bird image is available:
\dontrun{
check_docker()
}
}
