% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{check_docker}
\alias{check_docker}
\title{Check if Docker is running}
\usage{
check_docker(verbose = TRUE)
}
\arguments{
\item{verbose}{logical which indicates whether to print test results to
R console. On Windows always TRUE.}
}
\value{
If \code{verbose} is False: 0 upon success, otherwise an error code.
If \code{verbose} is True, no value is returned.
}
\description{
Checks that \href{https://www.docker.com/}{Docker} daemon is running
correctly on the local system
}
