% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{check_docker}
\alias{check_docker}
\title{Check if Docker is running}
\usage{
check_docker(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. When \code{TRUE}, messages are printed to R console.}
}
\value{
\itemize{
\item \code{0}: Success.
\item \code{1}: Error: Docker vol2bird image not available.
\item \code{2}: Error: Docker daemon not running.
\item \code{3}: Error: Docker daemon not found.
}
}
\description{
Checks if the \href{https://www.docker.com/}{Docker} daemon is running correctly
on the local system and that vol2bird Docker image is available.
}
\examples{
\dontrun{
# Check if Docker is running and vol2bird image is available
check_docker()
}
}
\seealso{
\itemize{
\item \code{\link[=update_docker]{update_docker()}}
}
}
