% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vpfiles.R
\name{check_url_existence}
\alias{check_url_existence}
\title{Check if a specific URL exists}
\usage{
check_url_existence(url)
}
\arguments{
\item{url}{Weblink (char) to a potentially existing webpage.}
}
\value{
z With a length > 1 if the URL is existing and downloading would be
possible.
}
\description{
Check if a specific URL exists
}
\keyword{internal}
