% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{checkDocker}
\alias{checkDocker}
\title{Checks that Docker is running}
\usage{
checkDocker(verbose = T)
}
\arguments{
\item{verbose}{logical which indicates whether to print test results to R console. On Windows always TRUE.}
}
\value{
0 upon success, otherwise an error code.
}
\description{
Checks that \href{https://www.docker.com/}{Docker} daemon is running correctly on the local system
}
