% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_mixture.R
\name{clean_mixture}
\alias{clean_mixture}
\title{Partition mixtures of animals using assumptions on airspeeds.}
\usage{
clean_mixture(x, u, v, U, V, slow = 1, fast = 12, drop_slow_component = TRUE)
}
\arguments{
\item{x}{a mixture animal density or linear reflectivity eta.}

\item{u}{the mixture's ground speed u component (west to east)}

\item{v}{the mixture's ground speed v component (south to north)}

\item{U}{the west to east wind component}

\item{V}{the south to north wind component}

\item{slow}{the slow component's airspeed}

\item{fast}{the fast component's airspeed}

\item{drop_slow_component}{when TRUE (default) output density, ground speed and
heading for fast component, when FALSE for slow component.}
}
\value{
a data.frame with corrected density or reflectivity x, ground speed (u, v)
and heading in clockwise degrees from north.
}
\description{
Partition mixtures of animals using assumptions on airspeeds.
}
\examples{
# drop the slow component (typically insects)
clean_mixture(100,-13,13,-7,6, fast=12, slow=1)
# drop the fast component (typically birds)
clean_mixture(100,-13,13,-7,6, fast=12, slow=1, drop_slow_component=FALSE)
}
