% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_mixture.R
\name{clean_mixture}
\alias{clean_mixture}
\alias{clean_mixture.default}
\alias{clean_mixture.vpts}
\alias{clean_mixture.vp}
\title{Partition mixtures of animals using assumptions on airspeeds.}
\usage{
clean_mixture(
  x,
  ...,
  slow = 1,
  fast = 8,
  drop_slow_component = TRUE,
  drop_missing = FALSE,
  keep_mixture = FALSE
)

\method{clean_mixture}{default}(
  eta,
  u,
  v,
  U,
  V,
  slow = 1,
  fast = 8,
  drop_slow_component = TRUE,
  drop_missing = FALSE,
  keep_mixture = FALSE
)

\method{clean_mixture}{vpts}(
  x,
  ...,
  slow = 1,
  fast = 8,
  drop_slow_component = TRUE,
  drop_missing = FALSE,
  keep_mixture = FALSE
)

\method{clean_mixture}{vp}(
  x,
  ...,
  slow = 1,
  fast = 8,
  drop_slow_component = TRUE,
  drop_missing = FALSE,
  keep_mixture = FALSE
)
}
\arguments{
\item{x}{a \code{vp} or \code{vpts} object.}

\item{...}{\code{eta}, \code{u}, \code{v}, \code{U}, \code{V} arguments, taken from object for \code{vp} or \code{vpts} class.}

\item{slow}{the slow component's airspeed}

\item{fast}{the fast component's airspeed}

\item{drop_slow_component}{when TRUE (default) output density, ground speed and
heading for fast component, when FALSE for slow component.}

\item{drop_missing}{Values \code{x} without an associated ground speed
and wind speed are set to NA when \code{TRUE}, or returned unaltered when \code{FALSE} (default).}

\item{keep_mixture}{When \code{TRUE} store original mixture reflectivity and speeds as
renamed quantities with \code{mixture_} prefix}

\item{eta}{a mixture animal density or linear reflectivity eta.}

\item{u}{the mixture's ground speed u component (west to east)}

\item{v}{the mixture's ground speed v component (south to north)}

\item{U}{the west to east wind component}

\item{V}{the south to north wind component}
}
\value{
a data.frame with corrected density or reflectivity x, ground speed (u, v)
and heading in clockwise degrees from north.
}
\description{
Partition mixtures of animals using assumptions on airspeeds.
}
\examples{
# drop the slow component (typically insects)
clean_mixture(100,-13,13,-7,6, fast=12, slow=1)
# drop the fast component (typically birds)
clean_mixture(100,-13,13,-7,6, fast=12, slow=1, drop_slow_component=FALSE)
}
