% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{cmt}
\alias{cmt}
\title{Cumulative migration traffic}
\usage{
cmt(x, alt.min = 0, alt.max = Inf, alpha = NA, interval.max = Inf)
}
\arguments{
\item{x}{A \code{vpts}.}

\item{interval.max}{Maximum time interval belonging to a single profile in
seconds. Traffic rates are set to zero at times \code{t} for which no
profiles can be found within the period \code{t-interval.max/2} to
\code{t+interval.max/2}.}
}
\value{
Atomic vector with (cumulative) migration traffic in number of
individuals / km.
}
\description{
Cumulative migration traffic is calculated as the cumulative sum of the
migration traffic within each time step of a time series. Cumulative
migration traffic gives the number of individuals that have passed per km
perpendicular to the migratory direction at the position of the radar as a
function of time from the start of time series within the specified altitude
band.
}
\examples{
# get example time series of vertical profiles:
data(example_vpts)
# print cumulative migration traffic to console:
cmt(example_vpts)
# plot cumulative migration traffic:
plot(cmt(example_vpts), type = "l", xlab = "time" , ylab = "CMT [birds/km]")
}
\seealso{
\code{\link{bioRad-deprecated}}.
}
\keyword{internal}
