% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_ppi.R
\name{composite_ppi}
\alias{composite_ppi}
\title{Make a composite of multiple plan position indicators (\code{ppi})}
\usage{
composite_ppi(x, param = "DBZH", cells.dim = c(100, 100))
}
\arguments{
\item{x}{A list of \code{ppi}.}

\item{param}{Scan paramater/quantity to composite.}

\item{cells.dim}{integer. Vector with number of cells in each spatial
dimension.}
}
\value{
A \code{\link[=summary.ppi]{ppi}}.
}
\description{
Combines multiple plan position indicators (\code{ppi}) into a single
\code{ppi}. Can be used to make a composite of \code{ppi}'s from multiple
radars.
}
\details{
The latitude/longitude of the returned \code{ppi} use the WGS84
datum.
}
\examples{
# load the example polar scan:
data(SCAN)
# to be written ...
}
