% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_legacy.R
\name{convert_legacy}
\alias{convert_legacy}
\alias{convert_legacy.vp}
\alias{convert_legacy.vpts}
\title{Convert legacy bioRad objects to current version}
\usage{
convert_legacy(x)

\method{convert_legacy}{vp}(x)

\method{convert_legacy}{vpts}(x)
}
\arguments{
\item{x}{A \code{vp}, or \code{vpts} object.}
}
\value{
An updated object of the same class as the input.
}
\description{
Convert legacy bioRad objects (vp, vpts) that have become
obsolete and make them compatible with the current bioRad version.
}
\examples{
# convert a vp object:
convert_legacy(example_vp)

# convert a vpts object:
convert_legacy(example_vpts)
}
