% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_basemap.R
\name{download_basemap}
\alias{download_basemap}
\title{Download a basemap for \code{map(ppi)}}
\usage{
download_basemap(
  x,
  verbose = TRUE,
  zoom,
  alpha = 1,
  source = "stamen",
  maptype = "terrain",
  ...
)
}
\arguments{
\item{x}{An object of class \code{ppi}.}

\item{verbose}{Logical, whether to print information to console.}

\item{zoom}{Zoom level (optional), see \link[ggmap]{get_map}. An integer
from 3 (continent) to 21 (building). By default the zoom level matching the
ppi extent is selected automatically.}

\item{alpha}{Transparency of the basemap (0-1).}

<<<<<<< HEAD
\item{source}{String identifying which map service should be used: "stamen" or "google".}
=======
\item{source}{String identifying which map service should be used: "stamen", "osm" or "google"}
>>>>>>> master

\item{maptype}{Type of basemap to plot. For Stamen Maps: "terrain",
"terrain-background", "terrain-labels", "terrain-lines", "toner",
"toner-2010", "toner-2011", "toner-background", "toner-hybrid",
"toner-labels", "toner-lines", "toner-lite", "watercolor". For Google
Maps: "terrain", "satellite", "roadmap", "hybrid"
  see \link[ggmap]{get_map} for options}

\item{...}{Arguments to pass to \link[ggmap]{get_map} function.}
}
\description{
Downloads a Stamen Maps, OpenStreetMap, Google Maps or Naver Map base layer
map using \link[ggmap]{get_map}.
}
\details{
To use Google Maps as \code{source}, you will have to register with Google,
enable billing and provide an API key to ggmap (see the
\href{https://github.com/dkahle/ggmap#attention}{ggmap README}). This
is due to Google changing its API requirements.
}
\details{
\link{download_basemap} depends on package \link[ggmap](https://github.com/dkahle/ggmap).
The latest development version includes important bugfixes that are not
yet available on CRAN, see the \href{https://github.com/dkahle/ggmap}{ggmap}
page for install instructions.

Google has [recently changed its API requirements](https://developers.google.com/maps/documentation/geocoding/usage-and-billing),
and __ggmap__ users are now required to register with Google.

1. Users must register with Google. You can do this at https://cloud.google.com/maps-platform/.
There is a fair bit of free use before you incur charges, and even then the charges
 are modest for light use. When checked in Feb 2020, Google provides a free $200 credit each month,
with map request costing $0.002. This amounts to 100,000 free map downloads per month.
2. Users must enable the APIs they intend to use with Google
3. Inside R, after loading the new version of ggmap, you’ll need provide
ggmap with your API key with \link[ggmap]{register_google}

See \link[ggmap]{get_googlemaps} and \href{https://github.com/dkahle/ggmap}{ggmap} pages for details.

Your API key is _private_ and unique to you, so be careful not to share it online,
for example in a GitHub issue or saving it in a shared R script file.
}
\examples{
# load an example scan:
data(example_scan)
# print summary info for the scan:
example_scan
# make ppi for the scan
ppi <- project_as_ppi(example_scan)
# grab a basemap that matches the extent of the ppi:
\dontrun{
basemap <- download_basemap(ppi)
# map the reflectivity quantity of the ppi onto the basemap:
map(ppi, map = basemap, param = "DBZH")
# download a different type of basemap, e.g. a gray-scale image:
# see get_map() in ggmap library for full documentation of options
basemap <- download_basemap(ppi, maptype = "toner-lite")
# map the radial velocities onto the line image:
map(ppi, map = basemap, param = "VRADH")
}
}
