% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pvolfiles.R
\name{download_pvolfiles}
\alias{download_pvolfiles}
\title{Download polar volume (\code{pvol}) files from the NEXRAD archive}
\usage{
download_pvolfiles(
  date_min,
  date_max,
  radar,
  directory = ".",
  overwrite = FALSE,
  bucket = "noaa-nexrad-level2"
)
}
\arguments{
\item{date_min}{character. YYYY-MM-DD start date of file selection. Days will
be ignored.}

\item{date_max}{character. YYYY-MM-DD end date of file selection. Days will
be ignored.}

\item{directory}{character. Path to local directory where files should be
downloaded and unzipped.}

\item{overwrite}{logical. TRUE for re-downloading and overwriting previously
downloaded files of the same names.}

\item{radars}{character (vector). 5-letter country/radar code(s)
(e.g. "bejab") of radars to include in file selection.}
}
\description{
Download a selection of polar volume (\code{pvol}) files from the
\href{https://registry.opendata.aws/noaa-nexrad/}{NEXRAD Level II archive data}.
}
