% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vpfiles.R
\name{download_vpfiles}
\alias{download_vpfiles}
\title{Download vertical profile files (\code{vp}) from the ENRAM data repository}
\usage{
download_vpfiles(date_min, date_max, radars, directory = ".",
  overwrite = FALSE)
}
\arguments{
\item{date_min}{character. ISO 8601 start date of data selection. Days will
be ignored.}

\item{date_max}{character. ISO 8601 end date of data selection. Days will be
ignored.}

\item{directory}{character. Path to local directory for downloading and
unzipping files.}

\item{overwrite}{logical. TRUE for redownloading and overwriting previously
downloaded files of the same names.}

\item{radar}{character (vector). 5-letter country/radar code(s)
(e.g. "bejab") of radars to include in data selection.}
}
\description{
Download and unzip a selection of vertical profile files (\code{vp}) from the
\href{http://enram.github.io/data-repository/}{ENRAM data repository}, where
these are stored as monthly zips per radar.
}
\examples{
my_path <- "~/my/directory/"
# Download data from radars "bejab" and "bewid", even if previously
# downloaded (overwrite = TRUE). Will successfully download 2016-10 files,
# but show 404 error for 2016-11 files (as these are not available).
\dontrun{
download_vpfiles(
  date_min = "2016-10-01",
  date_max = "2016-11-30",
  radar = c("bejab", "bewid"),
  directory = my_path,
  overwrite = TRUE
)
}
}
