% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vpfiles.R
\name{download_vpfiles}
\alias{download_vpfiles}
\title{Download vertical profile (\code{vp}) files from the ENRAM data repository}
\usage{
download_vpfiles(
  date_min,
  date_max,
  radars,
  directory = ".",
  overwrite = FALSE
)
}
\arguments{
\item{date_min}{Character. Start date of file selection, in \code{YYYY-MM-DD}
format. Days will be ignored.}

\item{date_max}{Character. End date of file selection, in \code{YYYY-MM-DD}
format. Days will be ignored.}

\item{radars}{Character (vector). 5-letter country/radar code(s) to include
in file selection.}

\item{directory}{Character. Path to local directory where files should be
downloaded and unzipped.}

\item{overwrite}{Logical. When \code{TRUE}, re-download and overwrite previously
downloaded files of the same names.}
}
\description{
Downloads and unzips a selection of vertical profile (\code{vp}) files from the
\href{http://enram.github.io/data-repository/}{ENRAM data repository}, where
these are stored as monthly zips per radar.
}
\examples{
# Download (and overwrite) data from radars "bejab" and "bewid".
# Will successfully download 2016-10 files, but show 404 error for
# 2016-11 files, since these are not available.
\dontrun{
dir.create("~/bioRad_tmp_files")
download_vpfiles(
  date_min = "2016-10-01",
  date_max = "2016-11-30",
  radar = c("bejab", "bewid"),
  directory = "~/bioRad_tmp_files",
  overwrite = TRUE
)
# Clean up
unlink("~/bioRad_tmp_files", recursive = TRUE)
}
}
\seealso{
\itemize{
\item \code{\link[=select_vpfiles]{select_vpfiles()}}
\item \code{\link[=read_vpfiles]{read_vpfiles()}}
}
}
