% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_scan}
\alias{example_scan}
\title{Example object of class \code{scan}}
\format{
An object of class \code{scan} of dimension 5 x 480 x 360.
}
\usage{
example_scan
}
\description{
Example of a \code{\link[=summary.scan]{scan}} object with name
\code{example_scan}.
}
\examples{
# get summary of example scan:
summary(example_scan)

# example_scan was created with:
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")
pvol <- read_pvolfile(pvolfile)
example_scan <- pvol$scans[[1]]
\dontrun{
# save(example_scan, file = "data/example_scan.rda")
}
}
\keyword{datasets}
