% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_vp}
\alias{example_vp}
\title{Vertical profile (\code{vp}) example}
\format{
An object of class \code{vp} with 25 rows and 16 columns.
}
\usage{
example_vp
}
\description{
Example object of class \code{\link[=summary.vp]{vp}} with name \code{example_vp}.
}
\examples{
# Reload example_vp from package (e.g. in case it was altered)
data(example_vp)

# Get summary info
example_vp

# example_vp was created with
\dontrun{
vpfile <- system.file("extdata", "profile.h5", package = "bioRad")
example_vp <- read_vpfiles(vpfile)
save(example_vp, file = "data/example_vp.rda")
}
}
\keyword{datasets}
