% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_vp}
\alias{example_vp}
\title{Vertical profile (\code{vp}) example}
\format{
An object of class \code{vp} with 25 rows and 16 columns.
}
\usage{
example_vp
}
\value{
An example object of type 'vp' which represents a vertical profile.
}
\description{
Example of a \code{\link[=summary.vp]{vp}} object with name \code{example_vp}.
}
\examples{
# Reload example_vp from package (e.g. in case it was altered)
data(example_vp)

# Get summary info
example_vp
}
\seealso{
\itemize{
\item \code{\link[=summary.vp]{summary.vp()}}
}
}
\keyword{datasets}
