% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvolfile.R
\name{get_odim_object_type}
\alias{get_odim_object_type}
\title{Check the ODIM data class of a polar volume file}
\usage{
get_odim_object_type(file)
}
\arguments{
\item{file}{Character. Name of the file to check.}
}
\value{
Character. \code{PVOL} for polar volume, \code{VP} for
vertical profile, otherwise \code{NA}.
}
\description{
Checks which data class is contained in ODIM HDF5 file.
}
\examples{
# Locate the polar volume example file:
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")

# Check the data type:
get_odim_object_type(pvolfile) # > "PVOL"
}
