% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{Identify \code{NaN} in a dataframe}
\usage{
\method{is.nan}{data.frame}(x)
}
\arguments{
\item{x}{A \code{data.frame} object.}
}
\value{
A matrix of the same dimension as \code{x}, with \code{TRUE}/\code{FALSE} values for
whether each cell in the original data frame is a number or not.
}
\description{
Identify cells with \code{NaN} (not a number) in a data frame. Improves on the
default\code{\link[base:is.finite]{base::is.nan()}} function, which only works on vectors, by allowing
data frames as input.
}
\keyword{internal}
