% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.nan.data.frame.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{Identify \code{NaN} in a dataframe.}
\usage{
\method{is.nan}{data.frame}(x)
}
\arguments{
\item{x}{A data frame to be tested.}
}
\value{
A matrix of the same dimension as \code{x}, with TRUE/FALSE values
for whether each cell in the original data frame is a number or not.
\code{NaN} means 'Not a Number'.
}
\description{
Identify cells with \code{NaN} in a data frame. Improve on the default
\code{is.nan()} function, which only works on vectors, by allowing
data frames as input.
}
\keyword{internal}
