% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{is.PVOLfile}
\alias{is.PVOLfile}
\title{Check whether file is a polar volume}
\usage{
is.PVOLfile(filename)
}
\arguments{
\item{filename}{A string containing a filename}
}
\value{
TRUE when \code{filename} is a polar volume in readable format, otherwise FALSE
}
\description{
Checker whether a file is a polar volume that can be read with package \pkg{bioRad}
}
\examples{
is.PVOLfile("/not/a/valid/PVOL/file.h5")   #> FALSE

}

