% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvolfile.R
\name{is.pvolfile}
\alias{is.pvolfile}
\title{Check if a local file is a polar volume (\code{pvol})}
\usage{
is.pvolfile(file, filename = NULL)
}
\arguments{
\item{file}{Character. Name of the file to check.}

\item{filename}{Deprecated, use \code{file} instead.}
}
\value{
\code{TRUE} when the file is a polar volume in readable format, otherwise
\code{FALSE}.
}
\description{
Checks whether a file is a polar volume in ODIM hdf5 format that can be read
with bioRad. Currently evaluates to \code{FALSE} for NEXRAD and IRIS RAW polar
volume files.
}
\examples{
# Locate the polar volume example file
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")

# Check that the file is an ODIM hdf5 polar volume
is.pvolfile(pvolfile) # > TRUE
}
