% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvolfile.R
\name{is.pvolfile}
\alias{is.pvolfile}
\title{Check if a file is a polar volume (\code{pvol})}
\usage{
is.pvolfile(file)
}
\arguments{
\item{file}{Character. Path of the file to check.}
}
\value{
\code{TRUE} when the file is a polar volume (\code{pvol}) in readable format,
otherwise \code{FALSE}.
}
\description{
Checks whether a file is a polar volume (\code{pvol}) in the ODIM hdf5 format that
can be read with bioRad. Currently evaluates to \code{FALSE} for NEXRAD and
IRIS RAW polar volume files.
}
\examples{
# Locate the polar volume example file
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")

# Check that the file is an ODIM hdf5 polar volume
is.pvolfile(pvolfile) # > TRUE
}
\seealso{
\itemize{
\item \code{\link[=read_pvolfile]{read_pvolfile()}}
\item \code{\link[=is.pvol]{is.pvol()}}
}
}
