% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvolfile.R
\name{is.pvolfile}
\alias{is.pvolfile}
\title{Check if a file is a polar volume (\code{pvol})}
\usage{
is.pvolfile(filename)
}
\arguments{
\item{filename}{A string containing a filename}
}
\value{
TRUE when \code{filename} is a polar volume in readable format, otherwise FALSE
}
\description{
Checker whether a file is a polar volume that can be read with package \pkg{bioRad}
}
\examples{
volume <- system.file("extdata", "volume.h5", package="bioRad")
is.pvolfile(volume)   #> TRUE

}
