% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpfile.R
\name{is.vpfile}
\alias{is.vpfile}
\title{Check if a local file is a vertical profile (\code{vp})}
\usage{
is.vpfile(file, filename = NULL)
}
\arguments{
\item{file}{A string containing a filename}

\item{filename}{Deprecated argument, use file instead.}
}
\value{
TRUE when \code{filename} is a vertical profile, otherwise FALSE
}
\description{
Checker whether a file is a vertical profile that can be read with
package \pkg{bioRad}
}
\examples{
profile <- system.file("extdata", "profile.h5", package = "bioRad")
is.vpfile(profile)   #> TRUE

}
