% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_vpts_aloft.R
\name{list_vpts_aloft}
\alias{list_vpts_aloft}
\title{List aloft urls for time series of vertical profiles (\code{vpts}) of radar
stations}
\usage{
list_vpts_aloft(
  date_min = NULL,
  date_max = NULL,
  radars = NULL,
  format = "csv",
  source = "baltrad"
)
}
\arguments{
\item{date_min}{The first date to return urls for. In the shape of
YYYY-MM-DD.}

\item{date_max}{The last date to return urls for. In the shape of YYYY-MM-DD.}

\item{radars}{A character vector of radar stations to return urls for.}

\item{format}{(Character) The format of archive urls to return, either csv or
hdf5. Currently only csv urls are supported.}

\item{source}{(Character) Either \code{baltrad} or \code{ecog-04003}}
}
\value{
A character vector of aloft urls
}
\description{
List aloft urls for time series of vertical profiles (\code{vpts}) of radar
stations
}
\examples{
list_vpts_aloft(radars = "bejab")
}
