% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\alias{map.ppi}
\title{Plot a plan position indicator (\code{ppi}) on a map}
\usage{
map(x, ...)

\method{map}{ppi}(
  x,
  map,
  param,
  alpha = 0.7,
  xlim,
  ylim,
  zlim = c(-20, 20),
  ratio,
  radar_size = 3,
  radar_color = "red",
  n_color = 1000,
  palette = NA,
  ...
)
}
\arguments{
\item{x}{A \code{ppi} object.}

\item{...}{Arguments passed to \code{\link[ggmap:ggmap]{ggmap::ggmap()}}.}

\item{map}{Basemap to use, result of a call to \code{\link[=download_basemap]{download_basemap()}}.}

\item{param}{Character. Scan parameter to plot, e.g. \code{DBZH} or \code{VRADH}. See
\code{\link[=summary.param]{summary.param()}} for commonly available parameters.}

\item{alpha}{Numeric. Transparency of the data, value between 0 and 1.}

\item{xlim}{Numeric vector of length 2. Range of x values (degrees longitude)
to plot.}

\item{ylim}{Numeric vector of length 2. Range of y values (degrees latitude)
to plot.}

\item{zlim}{Numeric vector of length 2. The range of values to plot.}

\item{ratio}{Numeric. Aspect ratio between x and y scale, by default
\eqn{1/cos(latitude radar * pi/180)}.}

\item{radar_size}{Numeric. Size of the symbol indicating the radar position.}

\item{radar_color}{Character. Color of the symbol indicating the radar
position.}

\item{n_color}{Numeric. Number of colors (>=1) to use in the palette.}

\item{palette}{Character vector. Hexadecimal color values defining the plot
color scale, e.g. output from \code{\link[viridisLite:viridis]{viridisLite::viridis()}}.}
}
\value{
A ggmap object (a classed raster object with a bounding
box attribute).
}
\description{
Plots a plan position indicator (\code{ppi}) on a base layer using
\code{\link[ggmap:ggmap]{ggmap::ggmap()}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{ppi}: Plot a \code{ppi} object on a map.
}}

\examples{
# Project a scan as a ppi
ppi <- project_as_ppi(example_scan)
\dontrun{
# Create a basemap that matches the extent of the ppi
basemap <- download_basemap(ppi, maptype = "toner-lite")

# Map the radial velocity of the ppi onto the basemap
map(ppi, map = basemap, param = "VRADH")

# Extend the plotting range of velocities, from -50 to 50 m/s
map(ppi, map = basemap, param = "VRADH", zlim = c(-50, 50))

# Map the reflectivity
map(ppi, map = basemap, param = "DBZH")

# Change the color palette to Viridis colors
map(ppi, map = basemap, param = "DBZH", palette = viridis::viridis(100), zlim=c(-10,10))

# Give the data more transparency
map(ppi, map = basemap, param = "DBZH", alpha = 0.3)

# Change the appearance of the symbol indicating the radar location
map(ppi, map = basemap, radar_size = 5, radar_color = "blue")

# Crop the map
map(ppi, map = basemap, xlim = c(12.4, 13.2), ylim = c(56, 56.5))
}
}
\seealso{
\itemize{
\item \code{\link[=download_basemap]{download_basemap()}}
\item \code{\link[=project_as_ppi]{project_as_ppi()}}
}
}
