% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic.R
\name{mt}
\alias{mt}
\title{Migration traffic}
\usage{
mt(x, alt.min = 0, alt.max = Inf, alpha = NA, interval.max = Inf)
}
\arguments{
\item{x}{A \code{vpts}.}

\item{interval.max}{Maximum time interval belonging to a single profile in
seconds. Traffic rates are set to zero at times \code{t} for which no
profiles can be found within the period \code{t-interval.max/2} to
\code{t+interval.max/2}.}
}
\value{
A numeric value equal to migration traffic in number of
individuals / km.
}
\description{
Total migration traffic, which is calculated by time-integration of
migration traffic rates. Migration traffic gives the number of individuals
that have passed per km perpendicular to the migratory direction at the
position of the radar for the full period of the time series within the
specified altitude band.
}
\examples{
# get example time series of vertical profiles:
data(example_vpts)
example_vpts
# total migration traffic in full altitude band:
\dontrun{mt(example_vpts)}
# total migration traffic in 0-1000 meter band:
\dontrun{mt(example_vpts, alt.min = 0, alt.max = 1000)}
}
\seealso{
\code{\link{bioRad-deprecated}}.
}
\keyword{internal}
