% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{mtr}
\alias{mtr}
\alias{mtr.vp}
\alias{mtr.vplist}
\alias{mtr.vpts}
\title{Migration traffic rate}
\usage{
mtr(x, alt.min, alt.max)

\method{mtr}{vp}(x, alt.min = 0, alt.max = Inf)

\method{mtr}{vplist}(x, alt.min = 0, alt.max = Inf)

\method{mtr}{vpts}(x, alt.min = 0, alt.max = Inf)
}
\arguments{
\item{x}{a \code{vp}, \code{vplist} or \code{vpts} object}

\item{alt.min}{minimum altitude in m}

\item{alt.max}{maximum altitude in m}
}
\value{
class \code{vp}: the migration traffic rate (MTR) individuals/km/h

class \code{vplist}: a numeric atomic vector with migration traffic rates in individuals/km/h

class \code{vpts}: a data frame with dates and migration traffic rates in individuals/km/h
}
\description{
Migration traffic rate (MTR) for an altitude layer, defined as the
number of targets crossing a 1 km line perpendicular to the migratory movement per hour
}
\section{Methods (by class)}{
\itemize{
\item \code{vp}: MTR of a vertical profile

\item \code{vplist}: MTR of a list of vertical profiles

\item \code{vpts}: MTR of a time series of vertical profiles
}}
\examples{
### MTR for a single vertical profile ###
mtr(VP)

### MTRs for a list of vertical profiles ###
mtr(c(VP,VP))

### MTRs for a time series of vertical profiles ###
# locate example file:
VPtable <- system.file("extdata", "VPtable.txt", package="bioRad")
# load time series:
ts=readvp.table(VPtable,radar="KBGM", wavelength='S')
# print migration traffic rates:
mtr(ts)
# plot migration traffic rates for the full air column:
plot(mtr(ts),type='l',xlab="time [UTC]",ylab="MTR [birds/km/h]")
#' plot migration traffic rates for altitudes > 1 km above sea level
plot(mtr(ts,alt.min=1000),type='l',xlab="time [UTC]",ylab="MTR [birds/km/h]")
}

