% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{mtr}
\alias{mtr}
\title{Migration traffic rate}
\usage{
mtr(x, alt.min = 0, alt.max = Inf)
}
\arguments{
\item{x}{a \code{vp}, \code{vplist} or \code{vpts} object}

\item{alt.min}{minimum altitude in m}

\item{alt.max}{maximum altitude in m}
}
\value{
an atomic vector of migration traffic rates in individuals/km/hour
}
\description{
Migration traffic rate (MTR) for an altitude layer, defined as the
number of targets crossing a 1 km line perpendicular to the migratory movement per hour
}
\details{
This is a wrapper function for \link[bioRad]{vintegrate}, extracting only the
migration traffic rate data.
}
\examples{
### MTR for a single vertical profile ###
mtr(VP)
### MTRs for a time series of vertical profiles ###
# load example time series
data(VPTS)
# print migration traffic rates:
mtr(VPTS)
# to plot migration traffic rate data, use vintegrate:
plot(vintegrate(VPTS),quantity="mtr")
}
