% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{MTR}
\alias{MTR}
\alias{MTR.VP}
\alias{MTR.VPList}
\alias{MTR.VPTimeSeries}
\title{Calculate migration traffic rate (MTR)}
\usage{
MTR(x, ...)

\method{MTR}{VP}(x, alt.min = 0, alt.max = Inf)

\method{MTR}{VPList}(x, alt.min = 0, alt.max = Inf)

\method{MTR}{VPTimeSeries}(x, alt.min = 0, alt.max = Inf)
}
\arguments{
\item{x}{a \code{VP}, \code{VPList} or \code{VPTimeSeries} object}

\item{alt.min}{minimum altitude in m}

\item{alt.max}{maximum altitude in m}
}
\value{
an atomic vector with MTRs in individuals/km/h
}
\description{
Calculate migration traffic rate (MTR) for an altitude layer, defined as the
number of targets crossing a 1 km line perpendicular to the migratory movement per hour
}
\section{Methods (by class)}{
\itemize{
\item \code{VP}: MTR of a vertical profile

\item \code{VPList}: MTR of a list of vertical profiles

\item \code{VPTimeSeries}: MTR of a list of vertical profiles
}}
\examples{
# MTR for a vertical profile:
MTR(VP)
# MTRs for a list of vertical profiles
MTR(c(VP,VP))
# MTRs for a time series:
ts=VPTimeSeries(c(VP,VP))
MTR(ts)
}

