% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexrad_to_odim.R
\name{nexrad_to_odim}
\alias{nexrad_to_odim}
\title{Convert a NEXRAD polar volume file to an ODIM HDF5 polar volume file}
\usage{
nexrad_to_odim(
  pvolfile_nexrad,
  pvolfile_odim,
  verbose = FALSE,
  mount = dirname(pvolfile_nexrad),
  local_install
)
}
\arguments{
\item{pvolfile_nexrad}{Character. Path to a polar volume file in NEXRAD/RSL
format.}

\item{pvolfile_odim}{Character. Path of the polar volume file in ODIM HDF5
format to be written to disk.}

\item{verbose}{Logical. When \code{TRUE}, Docker \code{stdout} is piped to the R
console. Always \code{TRUE} on Windows.}

\item{mount}{Character. Directory path of the mount point for the Docker
container.}

\item{local_install}{Character. Path to local vol2bird installation (e.g.
\code{your/vol2bird_install_directory/vol2bird/bin/vol2bird}) to use instead of
the Docker container.}
}
\value{
\code{TRUE} on successful conversion.
}
\description{
Converts a NEXRAD polar volume input file to an ODIM HDF5 polar volume output
file (see \code{\link[=is.pvolfile]{is.pvolfile()}}). Requires a running
\href{https://www.docker.com/}{Docker} daemon, see \code{\link[=check_docker]{check_docker()}}.
}
\examples{
\dontrun{
# Download a NEXRAD file and save as KBGM_example
download.file(
  "https://noaa-nexrad-level2.s3.amazonaws.com/2019/10/01/KBGM/KBGM20191001_000542_V06",
  "~/KBGM_example"
)

# Convert to ODIM format
nexrad_to_odim("~/KBGM_example", "~/KBGM_example.h5")

# Verify that we have generated a polar volume in ODIM HDF5 format
get_odim_object_type("~/KBGM_example.h5")

# Clean up
file.remove("~/KBGM_example", "~/KBGM_example.h5")
}
}
\seealso{
\itemize{
\item \code{\link[=check_docker]{check_docker()}}
\item \code{\link[=is.pvolfile]{is.pvolfile()}}
}
}
