% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyquist_velocity.R
\name{nyquist_velocity}
\alias{nyquist_velocity}
\title{Nyquist velocity for a given pulse repetition frequency (PRF)}
\usage{
nyquist_velocity(wavelength, prf1, prf2)
}
\arguments{
\item{wavelength}{radar wavelength in cm}

\item{prf1}{radar pulse repetition frequency in Hz}

\item{prf2}{alternate radar pulse repetition frequency in Hz (for a radar opering in dual-PRF mode)}
}
\value{
Nyquist velocity in m/s.
}
\description{
Calculates the Nyquist velocity given a radar's pulse repetition frequency (PRF)
and wavelength. When specifying two PRFs, the extended Nyquist
velocity is given for a radar using the dual-PRF technique.
}
\examples{
nyquist_velocity(5.3, 2000, 1500)
}
