% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.ppi.R
\name{plot.ppi}
\alias{plot.ppi}
\title{Plot a plan position indicator (PPI)}
\usage{
\method{plot}{ppi}(x, param, xlim, ylim, zlim = c(-20, 20), ratio = 1, ...)
}
\arguments{
\item{x}{an object of class 'ppi'}

\item{param}{the scan parameter to plot, see details below}

\item{xlim}{range of x values to plot}

\item{ylim}{range of y values to plot}

\item{zlim}{the range of parameter values to plot}

\item{ratio}{aspect ratio between x and y scale}

\item{...}{arguments passed to low level \link[ggplot2]{ggplot} function}
}
\description{
Plots a plan position indicator (PPI) generated with \link{ppi} using \link[ggplot2]{ggplot}
}
\details{
Available scan parameters for plotting can by printed to screen by \code{summary(x)}.
Commonly available parameters are:
\describe{
 \item{"\code{DBZH}", "\code{DBZ}"}{(Logged) reflectivity factor [dBZ]}
 \item{"\code{VRADH}", "\code{VRAD}"}{Radial velocity [m/s]. Radial velocities towards
  the radar are negative, while radial velocities away from the radar are positive}
 \item{"\code{RHOHV}"}{Correlation coefficient [unitless]. Correlation between vertically polarized and horizontally polarized reflectivity factor}
 \item{"\code{PHIDP}"}{Differential phase [degrees]}
 \item{"\code{ZDR}"}{(Logged) differential reflectivity [dB]}
}
The scan parameters are named according to the OPERA data information model (ODIM), see
Table 16 in the \href{http://www.eumetnet.eu/sites/default/files/OPERA2014_O4_ODIM_H5-v2.2.pdf}{ODIM specification}.
}
\examples{
# load an example scan:
data(SCAN)
# print to screen the available scan parameters:
summary(SCAN)
# make ppi for the scan
ppi=ppi(SCAN)
# plot the first scan parameter, which in this case is "VRADH":
plot(ppi)
# plot the reflectivity quantity:
plot(ppi,param="DBZH")
# change the range of reflectivities to plot to -30 to 50 dBZ:
plot(ppi,param="DBZH",zlim=c(-30,50))
}
