% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scan.R
\name{plot.scan}
\alias{plot.scan}
\title{Plot a scan (\code{scan}) in polar coordinates}
\usage{
\method{plot}{scan}(
  x,
  param,
  xlim = c(0, 1e+05),
  ylim = c(0, 360),
  zlim = c(-20, 20),
  na.value = "transparent",
  ...
)
}
\arguments{
\item{x}{A \code{scan} object.}

\item{param}{Character. Scan parameter to plot, e.g. \code{DBZH} or \code{VRADH}. See
\code{\link[=summary.param]{summary.param()}} for commonly available parameters.}

\item{xlim}{Numeric vector of length 2. Range of x values (range, distance to
radar) to plot.}

\item{ylim}{Numeric vector of length 2. Range of y values (azimuth) to plot.}

\item{zlim}{Numeric vector of length 2. The range of parameter values to
plot.}

\item{na.value}{Character. \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} parameter to set the color of
\code{NA} values.}

\item{...}{Arguments passed to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}
}
\description{
Plots a scan (\code{scan}) in polar coordinates. To plot in Cartesian coordinates,
see \code{\link[=project_to_ppi]{project_to_ppi()}}.
}
\examples{
# Plot reflectivity
plot(example_scan, param = "DBZH")
\dontrun{
# Change the range of reflectivities to plot, from -10 to 10 dBZ
plot(example_scan, param = "DBZH", zlim = c(-10, 10))

# Change the scale name, change the color palette to Viridis colors
plot(example_scan, param = "DBZH", zlim = c(-10, 10)) +
  viridis::scale_fill_viridis(name = "dBZ")
}
}
