% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{plot.vp}
\alias{plot.vp}
\title{Plot a vertical profile}
\usage{
\method{plot}{vp}(x, xlab = "density [#/km^3]", ylab = "height [km]",
  main = "Vertical profile", line.col = "red", line.lwd = 1, ...)
}
\arguments{
\item{x}{a vp class object}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{main}{an overall title for the plot}

\item{line.col}{Color of the plotted curve}

\item{line.lwd}{Line width of the plotted curve}

\item{...}{Additional arguments to be passed to the low level \link[graphics]{plot} plotting function}
}
\description{
Plot a vertical profile
}
\examples{
data(VP)
plot(VP)
plot(VP,line.col='blue')
}

