% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{print.SCAN}
\alias{print.SCAN}
\title{print method for class \code{SCAN}}
\usage{
\method{print}{SCAN}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{SCAN}, a polar scan}
}
\description{
print method for class \code{SCAN}
}
\keyword{internal}

