% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{print.scan}
\alias{print.scan}
\title{Print method for class \code{scan}}
\usage{
\method{print}{scan}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{scan}, a polar scan.}
}
\description{
Print method for class \code{scan}
}
\keyword{internal}
