% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{print.VP}
\alias{print.VP}
\title{print method for class \code{VP}}
\usage{
\method{print}{VP}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{VP}, usually a result of a call to \link[bioRad]{readVP}}
}
\description{
print method for class \code{VP}
}
\keyword{internal}

