% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp.r
\name{print.vp}
\alias{print.vp}
\title{Print method for class \code{vp}}
\usage{
\method{print}{vp}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{vp}, like the result of a call to
\link[=summary.vp]{read_vpfiles}.}
}
\description{
Print method for class \code{vp}
}
\keyword{internal}
