% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{print.vplist}
\alias{print.vplist}
\title{print method for class \code{vplist}}
\usage{
\method{print}{vplist}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{vplist}, usually a result of a call to \link{readvp.list}}
}
\description{
print method for class \code{vplist}
}
\keyword{internal}
