% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpts.R
\name{print.vpts}
\alias{print.vpts}
\title{print method for class \code{vpts}}
\usage{
\method{print}{vpts}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{vpts}, usually a result of a call to \link{vpts}}
}
\description{
print method for class \code{vpts}
}
\keyword{internal}
