% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_as_ppi.R
\name{proj_to_wgs}
\alias{proj_to_wgs}
\title{A wrapper for \code{\link[=spTransform]{spTransform()}}.}
\usage{
proj_to_wgs(x, y, proj4string)
}
\arguments{
\item{x}{Longitude}

\item{y}{Latitude}

\item{proj4string}{An object of class 'CRS', as defined in package \code{sp}.}
}
\value{
An object of class \code{SpatialPoints}.
}
\description{
A wrapper for \code{\link[=spTransform]{spTransform()}}.
}
\keyword{internal}
