% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_as_corrected_ppi.R
\name{project_as_corrected_ppi}
\alias{project_as_corrected_ppi}
\title{calculate a range-bias corrected PPI}
\usage{
project_as_corrected_ppi(pvol, vp, nx = 100, ny = 100, xlim = NA,
  ylim = NA, res = NA, param = "DBZH", lat = NA, lon = NA,
  antenna = NA, beam_angle = 1, crs = NA, quantity = c("vir",
  "vid", "correction_factor", "overlap"), k = 4/3, re = 6378,
  rp = 6357)
}
\arguments{
\item{pvol}{a polar volume of class pvol}

\item{vp}{a vertical profile of class vp}

\item{nx}{number of raster pixels in the x (longitude) dimension}

\item{ny}{number of raster pixels in the y (latitude) dimension}

\item{xlim}{x (longitude) range}

\item{ylim}{y (latitude) range}

\item{res}{numeric vector of length 1 or 2 to set the resolution of the raster (see \link[raster]{res}).
If this argument is used, arguments \code{nx} and \code{ny} are ignored. Unit is identical to \code{xlim} and \code{ylim}}

\item{param}{reflectivity factor scan parameter on which to base range bias corrections.
Typically the same parameter from which animal densities are estimated for object \code{vp}.
One of 'DBZH','DBZV','DBZ','TH','TV'.}

\item{lat}{Geodetic latitude of the radar in degrees. If \code{NA} taken from \code{pvol}.}

\item{lon}{Geodetic latitude of the radar in degrees. If \code{NA} taken from \code{pvol}.}

\item{beam_angle}{numeric. Beam opening angle in degrees, typically the
the angle between the half-power (-3 dB) points of the main lobe}

\item{crs}{character or object of class CRS. PROJ.4 type description of a Coordinate Reference System (map projection).
When 'NA' (default), an azimuthal equidistant projection with origin at the radar location is used.
To use a WSG84 (lat,lon) projection, use crs="+proj=longlat +datum=WGS84"}

\item{quantity}{one or multiple of 'vir','vid','correction_factor', 'overlap'}

\item{k}{Standard refraction coefficient.}

\item{re}{Earth equatorial radius in km.}

\item{rp}{Earth polar radius in km.}
}
\value{
An object of class '\link[=summary.ppi]{ppi}'.
}
\description{
calculates a PPI that corrects for range-bias effects due to partial beam overlap with
the layer of migration (overshooting) at larger distances from the radar
}
\details{
to be written
If one of \code{lat} or \code{lon} equals \code{NA}, the extent of the PPI is take equal to
the extent of the data in the first scan of the polar volume.

Overlap between vertical profile and vertical radiation profile is calculated using \link{beam_profile}
and stored as quantity 'overlap'.

TODO: quantity and param is used in a confusing way, correct.
}
\examples{
# locate example polar volume file:
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")
# load polar volume
example_pvol <- read_pvolfile(pvolfile)
# load the corresponding vertical profile for this polar volume
data(example_vp)
# calculate the range-bias corrected ppi on a 100x100 pixel raster
my_ppi <- project_as_corrected_ppi(example_pvol,example_vp,nx=100,ny=100)
# plot the vertically integrated reflectivity (vir) using a 0-2000 cm^2/km^2 color scale:
plot(my_ppi,zlim=c(0,2000))
# calculate the range-bias corrected ppi on finer 1000m x 1000m pixel raster:
my_ppi <- project_as_corrected_ppi(example_pvol,example_vp,res=1000)
# plot the vertically integrated density (vid) using a 0-200 birds/km^2 color scale:
plot(my_ppi,param="vid",zlim=c(0,200))
#' # download a basemap, and map the ppi:
bm = download_basemap(my_ppi)
map(my_ppi, bm)
# calculate the range-bias corrected ppi on an even finer 500m x 500m pixel raster,
# cropping the area up to 50000 meter from the radar.
my_ppi <- project_as_corrected_ppi(example_pvol,example_vp,res=500,xlim=c(-50000,50000),ylim=c(-50000,50000))
plot(my_ppi,param="vid",zlim=c(0,200))
}
