% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcs.R
\name{rcs<-}
\alias{rcs<-}
\alias{rcs<-.vp}
\alias{rcs<-.list}
\alias{rcs<-.vpts}
\alias{rcs<-.vpi}
\title{Change or set radar cross section}
\usage{
rcs(x) <- value

\method{rcs}{vp}(x) <- value

\method{rcs}{list}(x) <- value

\method{rcs}{vpts}(x) <- value

\method{rcs}{vpi}(x) <- value
}
\arguments{
\item{x}{A \code{vp}, list of \code{vp}, \code{vpts} or \code{vpi} object.}

\item{value}{Double. The radar cross section value to assign (cm^2).}
}
\description{
Changes or sets the assumed radar cross section in cm^2. This method also
updates the migration densities in \code{x$data$dens}.
}
\examples{
# Change or set radar cross section for a vp:
rcs(example_vp) <- 11

# Change or set radar cross section for a vpts:
rcs(example_vpts) <- 11

# Change or set radar cross section for a vpi:
example_vpi <- integrate_profile(example_vpts)
rcs(example_vpi) <- 11
}
\seealso{
\code{\link[=rcs]{rcs()}} for retrieving the radar cross section of an object.
}
