% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{rcs<-}
\alias{rcs<-}
\title{change radar cross section}
\usage{
rcs(x, ...) <- value
}
\arguments{
\item{x}{a \code{VP}, \code{VPList} or \code{VPTimeSeries} object}
}
\description{
change radar cross section
}
\examples{
# change RCS for a single vertical profile:
rcs(VP)<-20
}

