% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{rcs<-}
\alias{rcs<-}
\alias{rcs<-.vivp}
\alias{rcs<-.vp}
\alias{rcs<-.vplist}
\alias{rcs<-.vpts}
\title{Set radar cross section}
\usage{
rcs(x) <- value

\method{rcs}{vp}(x) <- value

\method{rcs}{vplist}(x) <- value

\method{rcs}{vpts}(x) <- value

\method{rcs}{vivp}(x) <- value
}
\arguments{
\item{x}{a \code{vp}, \code{vplist} or \code{vpts} object}

\item{value}{the cross section value to assign}
}
\description{
Sets the assumed radar cross section in cm^2. This method also updates the migration densities in \code{x$data$dens}
}
\examples{
# change RCS for a single vertical profile:
rcs(VP)<-20
}

