% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{rcs<-}
\alias{rcs<-}
\alias{rcs<-.VP}
\alias{rcs<-.VPList}
\alias{rcs<-.VPTimeSeries}
\title{Set radar cross section}
\usage{
rcs(x, ...) <- value

\method{rcs}{VP}(x) <- value

\method{rcs}{VPList}(x) <- value

\method{rcs}{VPTimeSeries}(x) <- value
}
\arguments{
\item{x}{a \code{VP}, \code{VPList} or \code{VPTimeSeries} object}
}
\description{
Sets the assumed radar cross section in cm^2. This method also updates the migration densities in \code{x$data$dens}
}
\examples{
# change RCS for a single vertical profile:
rcs(VP)<-20
}

