% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcs.R
\name{rcs}
\alias{rcs}
\alias{rcs.vp}
\alias{rcs.list}
\alias{rcs.vpts}
\alias{rcs.vpi}
\title{Get radar cross section}
\usage{
rcs(x)

\method{rcs}{vp}(x)

\method{rcs}{list}(x)

\method{rcs}{vpts}(x)

\method{rcs}{vpi}(x)
}
\arguments{
\item{x}{A \code{vp}, list of \code{vp}, \code{vpts} or \code{vpi} object.}
}
\value{
A radar cross section in cm^2.
}
\description{
Gives the currently assumed radar cross section in cm^2.
}
\section{Methods (by class)}{
\itemize{
\item \code{vpts}: Get radar cross section of a time series of vertical profile
(\code{vpts}).
}}

\examples{
# Retrieve radar cross section for a vp:
rcs(example_vp)

# Retrieve radar cross section for a vpts:
rcs(example_vpts)

# Retrieve radar cross section for a vpi:
example_vpi <- integrate_profile(example_vpts)
rcs(example_vpi)
}
\seealso{
\link{rcs()<-} for changing or setting the radar cross section of an
object.
}
