% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{rcs}
\alias{rcs}
\alias{rcs.vivp}
\alias{rcs.vp}
\alias{rcs.vplist}
\alias{rcs.vpts}
\title{Radar cross section}
\usage{
rcs(x)

\method{rcs}{vp}(x)

\method{rcs}{vplist}(x)

\method{rcs}{vpts}(x)

\method{rcs}{vivp}(x)
}
\arguments{
\item{x}{a \code{vp}, \code{vplist} or \code{vpts} object}
}
\value{
a radar cross section in cm^2
}
\description{
Gives the currently assumed radar cross section in cm^2.
}
\section{Methods (by class)}{
\itemize{
\item \code{vp}: radar cross section of a vertical profile

\item \code{vplist}: radar cross sections for a list of vertical profiles

\item \code{vpts}: radar cross section of a time series of vertical profile

\item \code{vivp}: radar cross section of a time series of vertically integrated vertical profile(s)
}}
\examples{
# extract RCS for a single vertical profile:
rcs(VP)
}

