% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{read.pvol}
\alias{read.pvol}
\title{Read a polar volume (PVOL) from file}
\usage{
read.pvol(filename, param = c("DBZH", "VRADH", "RHOHV", "ZDR", "PHIDP"),
  sort = T)
}
\arguments{
\item{filename}{A string containing the path to a vertical profile generated by \link[bioRad]{vol2bird}}

\item{sort}{logical. When \code{TRUE} sort scans ascending by elevation}

\item{params}{atomic vector of character strings, containing the names of scan parameters to read}
}
\value{
an object of class \code{PVOL}, which is a list containing polar scans, i.e. objects of class \code{SCAN}
\describe{
 \item{\strong{\code{radar}}}{the radar identifier}
 \item{\strong{\code{datetime}}}{the nominal time of the volume [UTC]}
 \item{\strong{\code{scans}}}{a list with scan objects}
 \item{\strong{\code{attributes}}}{list with the volume's \code{\\what}, \code{\\where} and \code{\\how} attributes}
}
}
\description{
Read a polar volume (PVOL) from file
}
\examples{
# locate example profile file:
pvol <- system.file("extdata", "volume.h5", package="bioRad")
# print the local path of the profile file:
pvol
# load the file:
read.pvol(prof)

}

