% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vpfiles.R
\name{read_vp}
\alias{read_vp}
\title{Read a vertical profile (\code{vp}) from file}
\usage{
read_vp(file)
}
\arguments{
\item{file}{A string containing the path to a vertical profile
generated by \link{calculate_vp}.}
}
\value{
An object of class \code{\link[=summary.vp]{summary.vp()}}.
}
\description{
Read a vertical profile (\code{vp}) from file
}
\keyword{internal}
