% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vpts.R
\name{read_vpts}
\alias{read_vpts}
\title{Read a time series of vertical profiles (\code{vpts}) from file}
\usage{
read_vpts(file, radar, wavelength = "C")
}
\arguments{
\item{file}{A text file containing the standard output (stdout) generated
by vol2bird (or the package function \code{calculate_vp}).}

\item{radar}{A string containing a radar identifier.}

\item{wavelength}{Radar wavelength in cm, or one of 'C' or 'S' for C-band
and S-band radar, respectively.}
}
\value{
An object inhereting from class \code{vpts}, see
\code{\link[=summary.vpts]{vpts}} for details.
}
\description{
Read a time series of vertical profiles (\code{vpts}) from file
}
\examples{
# locate example file:
vptszipfile <- system.file("extdata", "vpts.txt.zip", package = "bioRad")
\dontrun{
unzip(vptszipfile,"your/directory/and/file/name.txt")
# load time series:
ts <- read_vpts("your/directory/and/file/name.txt", radar = "KBGM", wavelength = "S")
ts
}
}
