% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vpts.R
\name{read_vpts}
\alias{read_vpts}
\title{Read time series of vertical profiles (\code{vpts}) from file(s)}
\usage{
read_vpts(files, ...)
}
\arguments{
\item{files}{Path(s) to one or more files containing vpts data.}

\item{...}{Additional arguments for backward compatibility, passed to \code{read_stdout}.}
}
\value{
\code{vpts} object.
}
\description{
Reads \code{vpts} data from one or more files.
The following file formats are supported (but cannot be mixed):
\itemize{
\item \href{https://aloftdata.eu/vpts-csv/}{VPTS CSV}.
\item \href{https://github.com/adokter/vol2bird/wiki/ODIM-bird-profile-format-specification}{ODIM bird profile}.
}
}
\concept{read functions}
