% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{readVP.list}
\alias{readVP.list}
\title{Read a list of vertical profiles from multiple files}
\usage{
readVP.list(files)
}
\arguments{
\item{files}{A string vector containing the filenames of vertical profiles in ODIM HDF5 format generated by \link[bioRad]{vol2bird}}
}
\value{
an object of class \code{VP}, which is a list containing
\describe{
 \item{\code{VPList}}{vertical profiles, a list of \code{VP} objects, see \link[bioRad]{readVP} for a description}
 \item{\code{radars}}{the radars represented in the list}
 \item{\code{daterange}}{the minimum and maximum nominal time of the profiles in the list}
 \item{\code{dates}}{the nominal times of the profiles}
}
}
\description{
Read a list of vertical profiles from multiple files
}
\examples{
\dontrun{readVP(c("my/path/profile1.h5","my/path/profile2.h5", ...))}

}

