% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{readvp}
\alias{readvp}
\title{Read a vertical profile (vp) from file}
\usage{
readvp(filename)
}
\arguments{
\item{filename}{A string containing the path to a vertical profile generated by \link[bioRad]{vol2bird}}
}
\value{
an object of class \link[bioRad]{summary.vp}
}
\description{
Read a vertical profile (vp) from file
}
\examples{
# locate example profile file:
prof <- system.file("extdata", "profile.h5", package="bioRad")
# print the local path of the profile file:
prof
# load the file:
readvp(prof)

}
