% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{regularize}
\alias{regularize}
\title{Regularize a time series}
\usage{
regularize(ts, interval = "auto", units = "mins", fill = F)
}
\arguments{
\item{ts}{an object inhereting from class \code{VPTimeSeries}, see \link[bioRad]{VPTimeSeries} for details}

\item{units}{optional units of \code{interval}, one of 'secs', 'mins', 'hours','days', 'weeks'. Defaults to 'mins'.}

\item{fill}{logical. Whether to fill empty timesteps with the values of the closest neighbouring profile}
}
\value{
an object of class \code{VPTimeSeries} with regular time steps
}
\description{
Projects objects of class \code{VPTimeSeries} on a regular time grid
}
\examples{
# locate example file:
VPtable <- system.file("extdata", "VPtable.txt", package="bioRad")
# load time series:
ts=readVP.table(VPtable,radar="KBGM")
# regularize the time series on a 5 minute interval grid
tsRegular=regularize(ts, interval=5)
}

