% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvsd.R
\name{rvsd<-}
\alias{rvsd<-}
\alias{rvsd<-.vp}
\alias{rvsd<-.vplist}
\alias{rvsd<-.vpts}
\title{Set threshold of the VVP-retrieved radial velocity standard deviation}
\usage{
rvsd(x) <- value

\method{rvsd}{vp}(x) <- value

\method{rvsd}{vplist}(x) <- value

\method{rvsd}{vpts}(x) <- value
}
\arguments{
\item{x}{a \code{vp}, \code{vplist} or \code{vpts} object}

\item{value}{the value to assign}
}
\description{
Sets the threshold in \code{sd_vvp}. Altitude layers with \code{sd_vvp}
below this threshold are assumed to have an aerial density of zero
individuals. This method updates the migration densities
in \code{x$data$dens}
}
\examples{
# change threshold for a single vertical profile:
rvsd(example_vp) <- 2
}
