% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{sd_vvp<-}
\alias{sd_vvp<-}
\alias{sd_vvp<-.VP}
\alias{sd_vvp<-.VPList}
\alias{sd_vvp<-.VPTimeSeries}
\title{Set threshold for VVP-retrieved radial velocity standard deviation}
\usage{
sd_vvp(x, ...) <- value

\method{sd_vvp}{VP}(x) <- value

\method{sd_vvp}{VPList}(x) <- value

\method{sd_vvp}{VPTimeSeries}(x) <- value
}
\arguments{
\item{x}{a \code{VP}, \code{VPList} or \code{VPTimeSeries} object}
}
\description{
Sets the threshold in \code{sd_vvp}. Altitude layers with \code{sd_vvp} below this threshold are
assumed to have an aerial density of zero individuals. This method updates the migration densities in \code{x$data$dens}
}
\examples{
# change threshold for a single vertical profile:
sd_vvp(VP)<-2
}

