% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{sd_vvp}
\alias{sd_vvp}
\alias{sd_vvp.vp}
\alias{sd_vvp.vplist}
\alias{sd_vvp.vpts}
\title{threshold VVP-retrieved radial velocity standard deviation}
\usage{
sd_vvp(x)

\method{sd_vvp}{vp}(x)

\method{sd_vvp}{vplist}(x)

\method{sd_vvp}{vpts}(x)
}
\arguments{
\item{x}{a \code{vp}, \code{vplist} or \code{vpts} object}
}
\value{
threshold for \code{sd_vvp} in m/s.
}
\description{
Gives the current threshold in VVP-retrieved radial velocity standard deviation in m/s.
}
\section{Methods (by class)}{
\itemize{
\item \code{vp}: threshold in VVP-retrieved radial velocity standard deviation of a vertical profile

\item \code{vplist}: threshold in VVP-retrieved radial velocity standard deviation of a list of vertical profiles

\item \code{vpts}: threshold in VVP-retrieved radial velocity standard deviation of a time series of vertical profiles
}}

\examples{
# extract threshold for a single vertical profile:
sd_vvp(VP)
}
