% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_vvp_threshold.R
\name{sd_vvp_threshold<-}
\alias{sd_vvp_threshold<-}
\alias{sd_vvp_threshold<-.vp}
\alias{sd_vvp_threshold<-.list}
\alias{sd_vvp_threshold<-.vpts}
\title{Set threshold of the radial velocity standard deviation}
\usage{
sd_vvp_threshold(x) <- value

\method{sd_vvp_threshold}{vp}(x) <- value

\method{sd_vvp_threshold}{list}(x) <- value

\method{sd_vvp_threshold}{vpts}(x) <- value
}
\arguments{
\item{x}{A \code{vp}, list of \code{vp} or \code{vpts} object.}

\item{value}{Double. The \code{sd_vvp} threshold value to assign in m/s.}
}
\description{
Sets the threshold of radial velocity standard deviation (\code{sd_vvp}) of an
object in m/s. Altitude layers with \code{sd_vvp} below this threshold are assumed
to have an aerial density of zero individuals. This method also updates the
migration densities in \code{x$data$dens}.
}
\examples{
# Set the sd_vvp threshold for a vp
sd_vvp_threshold(example_vp) <- 2

# Set the sd_vvp threshold for a vpts
sd_vvp_threshold(example_vp) <- 2
}
\seealso{
\code{\link[=sd_vvp_threshold]{sd_vvp_threshold()}} for getting the \code{sd_vvp} threshold of an
object.
}
