% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_vvp_threshold.R
\name{sd_vvp_threshold}
\alias{sd_vvp_threshold}
\alias{sd_vvp_threshold.vp}
\alias{sd_vvp_threshold.list}
\alias{sd_vvp_threshold.vpts}
\title{Get threshold of the VVP-retrieved radial velocity standard deviation}
\usage{
sd_vvp_threshold(x)

\method{sd_vvp_threshold}{vp}(x)

\method{sd_vvp_threshold}{list}(x)

\method{sd_vvp_threshold}{vpts}(x)
}
\arguments{
\item{x}{A \code{vp}, list of \code{vp} or \code{vpts} object.}
}
\value{
Threshold for `sd_vvp`` in m/s.
}
\description{
Gives the current threshold in VVP-retrieved radial velocity standard
deviation in m/s.
}
\examples{
# load example data:
data(example_vp)
data(example_vpts)

# retrieve threshold for a single vertical profile:
sd_vvp_threshold(example_vp)

# retrieve threshold for a vertical profile time series:
sd_vvp_threshold(example_vpts)

# change or set the threshold for a single vertical profile:
sd_vvp_threshold(example_vp) <- 2

# change or set the threshold for a vertical profile time series:
sd_vvp_threshold(example_vp) <- 2
}
\seealso{
\code{\link{sd_vvp_threshold()<-}} for setting an objects radial velocity
standard deviation.
}
