% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vpfiles.R
\name{select_vpfiles}
\alias{select_vpfiles}
\title{Select vertical profiles files (\code{vp}) from computer}
\usage{
select_vpfiles(path, date_min, date_max, country = NULL, radar = NULL)
}
\arguments{
\item{path}{Main path to look into recusively.}

\item{date_min}{ISO format date as start of the vp file query.}

\item{date_max}{ISO format date as end of the vp file query.}

\item{country}{Character vector with two letter country shortcuts.}

\item{radar}{Character vector with three letter radar sindicators. This
can be defined independently from the countries named.}
}
\value{
Character list of filenames that comply to the given
radar/country and date range query
}
\description{
Collect a list of vp file names within a directory that comply to the given
country, radar and date range combination
}
\examples{
my_path <- "~/my/directory/"
select_vpfiles(my_path, "2016-10-01", "2017-01-31", c("be"))
}
