% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.R
\name{[.ppi}
\alias{[.ppi}
\title{Subset a plan position indicator (\code{ppi})}
\usage{
\method{[}{ppi}(x, i)
}
\arguments{
\item{x}{An object of class \code{param} or \code{scan}.}

\item{i}{Indices specifying elements to extract.}
}
\description{
Select quantities by index from a \code{ppi}
}
\examples{
# make a ppi:
my_ppi <- project_as_ppi(example_scan)

# this ppi contains 5 quantities (VRADH DBZH ZDR RHOHV PHIDP):
my_ppi

# This ppi only contains the first quantity (VRADH):
my_ppi[1]

# This ppi contains the first three quantities (VRADH, DBZH, ZDR):
my_ppi[1:3]
}
