% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi.R
\name{[.ppi}
\alias{[.ppi}
\title{Subset a plan position indicator (\code{ppi})}
\usage{
\method{[}{ppi}(x, i)
}
\arguments{
\item{x}{An object of class \code{param} or \code{scan}.}

\item{i}{Indices specifying elements to extract.}
}
\description{
Extract by index from a ppi
}
