% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.ppi.R
\name{[.ppi}
\alias{[.ppi}
\title{Subset `ppi`}
\usage{
\method{[}{ppi}(x, i)
}
\arguments{
\item{x}{an object of class 'param' or 'scan'}

\item{i}{indices specifying elements to extract}
}
\description{
Extract by index from a ppi
}
