% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{[.vpts}
\alias{[.vpts}
\title{Subset `vpts`}
\usage{
\method{[}{vpts}(x, i)
}
\arguments{
\item{x}{object of class 'vpts'}

\item{i}{indices specifying elements to extract}
}
\description{
Extract by index from a vpts
}
