% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpts.R
\name{[.vpts}
\alias{[.vpts}
\title{Subset a time series of vertical profiles (\code{vpts})}
\usage{
\method{[}{vpts}(x, i)
}
\arguments{
\item{x}{A \code{vpts} object.}

\item{i}{Integer. Index/indices specifying which range of vertical profiles
to extract.}
}
\description{
Select a vertical profile (\code{vp}) or a time series of vertical profiles
(\code{vpts}) by index from a \code{vpts}.
}
\examples{
# Load the example time series of vertical profiles
data(example_vpts)

# This vpts contains 1934 profiles (i.e. datetimes)
dim(example_vpts)

# Subset vpts to extract 10th profile (returns a vp object)
example_vpts[10]

# Subset vpts to extract the 20th to 100th profile (returns a vpts object)
example_vpts[20:100]
}
