% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.ppi.R
\name{summary.ppi}
\alias{dim.ppi}
\alias{is.ppi}
\alias{summary.ppi}
\title{Class 'ppi': plan position indicator}
\usage{
\method{summary}{ppi}(object, ...)

is.ppi(x)

\method{dim}{ppi}(x)
}
\arguments{
\item{object}{object of class 'ppi'}

\item{...}{additional arguments affecting the summary produced.}

\item{x}{object of class 'ppi'}
}
\value{
for \code{is.ppi}: \code{TRUE} if its argument is of class "\code{ppi}"

for \code{dim.ppi}: dimensions of the ppi
}
\description{
Class 'ppi': plan position indicator
}
\details{
A PPI of class 'ppi' is a list containing:
\describe{
 \item{\code{data}}{an object of class \link[sp]{SpatialGridDataFrame} containing the georeferenced data}
 \item{\code{geo}}{geographic data, a list with:
    \describe{
     \item{\code{lat}}{latitude of the radar [decimal degrees]}
     \item{\code{lon}}{longitude of the radar [decimal degrees]}
     \item{\code{height}}{height of the radar antenna [metres above sea level]}
     \item{\code{elangle}}{radar beam elevation [degrees]}
     \item{\code{rscale}}{range bin size [m]}
     \item{\code{ascale}}{azimuth bin size [deg]}
    }
    The \code{geo} element of a 'scan' object is a copy of the \code{geo} element of its parent scan or scan parameter.
  }
}
}

