% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.ppi.R
\name{summary.pvol}
\alias{is.pvol}
\alias{summary.pvol}
\title{Class 'pvol': polar volume}
\usage{
\method{summary}{pvol}(object, ...)

is.pvol(x)
}
\arguments{
\item{object}{object of class 'pvol'}

\item{...}{additional arguments affecting the summary produced.}

\item{x}{object of class 'pvol'}
}
\value{
for \code{is.pvol}: \code{TRUE} if its argument is of class "\code{pvol}"
}
\description{
Class 'pvol': polar volume
}
\details{
A polar scan object of class 'pvol' is a list containing:
\describe{
 \item{\code{radar}}{character string with the radar identifier}
 \item{\code{datetime}}{nominal time of the volume [UTC]}
 \item{\code{scans}}{a list with scan objects of class 'scan'}
 \item{\code{attributes}}{list with the volume's \code{\\what}, \code{\\where} and \code{\\how} attributes}
 \item{\code{geo}}{geographic data, a list with:
  \describe{
     \item{\code{lat}}{latitude of the radar [decimal degrees]}
     \item{\code{lon}}{longitude of the radar [decimal degrees]}
     \item{\code{height}}{height of the radar antenna [metres above sea level]}
  }
 }
}
}
\examples{
is.pvol("this is not a polar volume but a string")  #> FALSE
}

