% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvol.R
\name{summary.pvol}
\alias{summary.pvol}
\alias{is.pvol}
\alias{print.pvol}
\title{Inspect a polar volume (\code{pvol})}
\usage{
\method{summary}{pvol}(x, ...)

is.pvol(x)

\method{print}{pvol}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{A \code{pvol} object.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
For \code{is.pvol(x)}: \code{TRUE} if \code{x} is an object of class \code{pvol},
otherwise \code{FALSE}.
}
\description{
R base functions for inspecting a polar volume (\code{pvol}) object.
}
\details{
A polar volume (\code{pvol}) object is a list containing:
\itemize{
\item \code{radar}: Radar identifier.
\item \code{datetime}: Nominal time of the volume (UTC).
\item \code{scans}: List of scans (\code{scan}) at different elevation angles.
\item \code{attributes}: List of the the volume's \code{what}, \code{where} and \code{how}
attributes.
\item \code{geo}: List of the volume's geographic properties:
\itemize{
\item \code{lat}: Latitude of the radar (decimal degrees).
\item \code{lon}: Longitude of the radar (decimal degrees).
\item \code{height}: Height of the radar antenna (meters above sea level).
}
}
}
\examples{
# Locate the polar volume example file:
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")

# Read the file:
example_pvol <- read_pvolfile(pvolfile)

# Verify that it is a pvol object:
is.pvol(example_pvol)

# Print summary info:
example_pvol # Same as summary(example_pvol) or print(example_pvol)

# Print summary info for the scans in the polar volume:
example_pvol$scans

# Copy the first scan to a new object `scan`:
scan <- example_pvol$scans[[1]]

# Print summary info for the scan:
scan # Same as summary(scan) or print(scan)
}
\seealso{
\itemize{
\item \code{\link[=read_pvolfile]{read_pvolfile()}} to read polar volume files.
\item \code{\link[=summary.scan]{summary.scan()}} to inspect scans.
}
}
