% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvol.R
\name{summary.pvol}
\alias{summary.pvol}
\alias{print.pvol}
\alias{is.pvol}
\title{Inspect a polar volume (\code{pvol})}
\usage{
\method{summary}{pvol}(x, ...)

\method{print}{pvol}(x, digits = max(3L, getOption("digits") - 3L), ...)

is.pvol(x)
}
\arguments{
\item{x}{A \code{pvol} object.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
For \code{\link[=is.pvol]{is.pvol()}}: \code{TRUE} for an object of class \code{pvol}, otherwise
\code{FALSE}.
}
\description{
R base functions for inspecting a polar volume (\code{pvol}) object.
}
\details{
A polar volume consists of a number of scans (or sweeps) made by the radar at
different elevation angles. A polar volume (\code{pvol}) object is a list
containing:
\itemize{
\item \code{radar}: Radar identifier.
\item \code{datetime}: Nominal time of the volume in UTC.
\item \code{scans}: List of scans (\code{scan}) at different elevation angles.
\item \code{attributes}: List of the volume's \code{what}, \code{where} and \code{how}
attributes.
\item \code{geo}: List of the volume's geographic properties:
\itemize{
\item \code{lat}: Latitude of the radar in decimal degrees.
\item \code{lon}: Longitude of the radar in decimal degrees.
\item \code{height}: Height of the radar antenna in meters above sea level.
}
}
}
\examples{
# Locate and read the polar volume example file
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")
example_pvol <- read_pvolfile(pvolfile)

# Verify that is an object of class pvol
is.pvol(example_pvol)

# Get summary info
example_pvol # Same as summary(example_pvol) or print(example_pvol)

# Get summary info for the scans in the polar volume
example_pvol$scans
}
\seealso{
\itemize{
\item \code{\link[=read_pvolfile]{read_pvolfile()}}
\item \code{\link[=get_elevation_angles]{get_elevation_angles()}}
\item \code{\link[=get_scan]{get_scan()}}
}
}
