% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{summary.VP}
\alias{dim.VP}
\alias{is.VP}
\alias{summary.VP}
\title{Class VP}
\usage{
\method{summary}{VP}(x)

is.VP(x)

\method{dim}{VP}(x)
}
\arguments{
\item{x}{object}
}
\value{
for \code{is.VP}: \code{TRUE} if its argument is of class "\code{VP}"

for \code{dim.VP}: dimensions of the profile data
}
\description{
Class for vertical profiles
}
\details{
details to be written
}

