% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{summary.VPList}
\alias{[.VPList}
\alias{[[.VPList}
\alias{is.VPList}
\alias{summary.VPList}
\title{Class VPList}
\usage{
\method{summary}{VPList}(x)

is.VPList(x)

\method{[}{VPList}(x, i)

\method{[[}{VPList}(x, i)
}
\arguments{
\item{x}{object}
}
\value{
for \code{is.VPList}: \code{TRUE} if its argument is of class "\code{VPList}"
}
\description{
Class for list of vertical profiles
}
\details{
details to be written
}
\keyword{internal}

