% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{summary.vplist}
\alias{summary.vplist}
\alias{[.vplist}
\alias{is.vplist}
\title{Class 'vplist': list of vertical profiles}
\usage{
\method{summary}{vplist}(object, ...)

\method{[}{vplist}(x, i)

is.vplist(x)
}
\arguments{
\item{object}{object of class 'vplist'}

\item{...}{additional arguments affecting the summary produced.}

\item{x}{object of class 'vplist'}

\item{i}{indices specifying elements to extract}
}
\value{
for \code{is.vplist}: \code{TRUE} if its argument is of class "\code{vplist}"
}
\description{
Class for list of vertical profiles
}
\details{
An object of class \code{vplist} is a list containing only \link[=summary.vp]{vp} objects.
By contrast, \link[=summary.vpts]{vpts} objects contain time-ordered profiles of a single radar
station. \code{vplist} objects can contain profiles of multiple radars.
}
