% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{summary.vpts}
\alias{[.vpts}
\alias{dim.vpts}
\alias{is.vpts}
\alias{summary.vpts}
\title{Class 'vpts': time series of vertical profiles}
\usage{
\method{summary}{vpts}(object, ...)

is.vpts(x)

\method{dim}{vpts}(x)

\method{[}{vpts}(x, i)
}
\arguments{
\item{object}{object of class 'vpts'}

\item{...}{additional arguments affecting the summary produced.}

\item{x}{object of class 'vpts'}

\item{i}{indices specifying elements to extract}
}
\value{
for \code{is.vpts}: \code{TRUE} if its argument is of class "\code{vpts}"

for \code{dim.vpts}: dimensions of the time series
}
\description{
Class for single-site time series of vertical profiles
}
\details{
details to be written
}

