% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunrise_sunset.R
\name{sunrise_sunset}
\alias{sunrise_sunset}
\alias{sunrise}
\alias{sunset}
\title{Calculate sunrise or sunset for a time and place}
\usage{
sunrise(date, lon, lat, elev = -0.268, tz = "UTC", force_tz = FALSE)

sunset(date, lon, lat, elev = -0.268, tz = "UTC", force_tz = FALSE)
}
\arguments{
\item{date}{Date inheriting from class \code{POSIXt} or a string
interpretable by \link[base]{as.Date}.}

\item{lon}{Longitude in decimal degrees.}

\item{lat}{Latitude in decimal degrees.}

\item{elev}{Sun elevation in degrees.}

\item{tz}{output time zone. Ignored if \code{date} has an associated time zone already}

\item{force_tz}{whether to convert input time to timezone \code{tz}. Default \code{FALSE}.}
}
\value{
The moment of sunrise or sunset for the date set by \code{date}and time zone as specified
(by \code{date} and \code{tz}) or in UTC if not specified.
}
\description{
Calculate sunrise or sunset for a time and place
}
\details{
The angular diameter of the sun is about 0.536 degrees,
therefore the moment of sunrise/sunset corresponds to half that elevation
at -0.268 degrees.

This is a convenience function mapping to \link{crepuscule}.

Approximate astronomical formula are used, therefore the moment of
sunrise / sunset may be off by a few minutes

Sunrise and sunset is given for the day specified by \code{date}, with sunset always
later than sunrise given the same \code{date}.

If \code{force_tz} is \code{TRUE}, the input date is converted to the timezone
set by \code{tz} prior to calculating the sunrise/sunset time, resulting in a shift by +/-1 day
}
\examples{
# sunrise in the Netherlands
sunrise("2016-01-01", 5, 53)

# sunset in the Netherlands
sunset("2016-01-01", 5, 53)

# civil twilight in Ithaca, NY, today
sunrise(Sys.time(), -76.5, 42.4, elev = -6)
}
