% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{update_docker}
\alias{update_docker}
\title{Update Docker image from Docker hub}
\usage{
update_docker(mistnet = FALSE)
}
\arguments{
\item{mistnet}{Logical. When \code{TRUE}, installs the MistNet segmentation model,
downloading an additional 1GB Docker image (see \code{\link[=apply_mistnet]{apply_mistnet()}}).}
}
\value{
The POSIXct creation date of the installed Docker image.
}
\description{
Pulls and installs the latest \href{https://hub.docker.com/r/adokter/vol2bird/}{vol2bird Docker image} from Docker hub. Run this
command to ensure all Docker functionality (e.g. for \code{\link[=calculate_vp]{calculate_vp()}}) makes
use of the latest available version.
}
\examples{
\dontrun{
# Update the vol2bird docker image
update_docker()

# Update the vol2bird docker image and install the MistNet segmentation model
update_docker(mistnet = TRUE)
}
}
\seealso{
\itemize{
\item \code{\link[=check_docker]{check_docker()}}
}
}
