% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vad.R
\name{vad}
\alias{vad}
\alias{vad.pvol}
\title{Create a velocity azimuth display (VAD) plot}
\usage{
vad(x, ...)

\method{vad}{pvol}(
  x,
  vp = NULL,
  ...,
  range = NULL,
  height = NULL,
  range_gate_filter = DBZH < 20 & RHOHV < 0.95,
  point_geom = ggplot2::geom_point,
  point_geom_args = list(),
  annotate = "{round(ff,1)} m/s, {round(dd)}°"
)
}
\arguments{
\item{x}{A polar volume from which range gates are extracted.}

\item{...}{Currently not used.}

\item{vp}{A vertical profile to annotate the VAD plot with the fit in the vertical profile}

\item{range}{The distance range in to filter the range gates with.
If a \code{vp} is provided the range is taken from there and the argument \code{range} should not be provided.}

\item{height}{The height range to filter the range gates by.
If only one value is provided next to a \code{vp} then the height bin intersection with this elevation is plotted.}

\item{range_gate_filter}{Optional filtering of the range gates. By default range gates are filtered for a \code{DBZH} less then 20 and a \code{RHOHV} less then 0.95.
Alternative filters could be used to highlight specific effects.}

\item{point_geom}{The geom to visualize the range gates, the default is \code{\link[=geom_points]{geom_points()}}, in some cases this suffers from over plotting.
Alternatives that avoid over plotting could be \code{\link[=geom_bin2d]{geom_bin2d()}} or \code{\link[ggpointdensity:geom_pointdensity]{ggpointdensity::geom_pointdensity()}}.}

\item{point_geom_args}{Additional arguments to the \code{point_geom} function. For example controling the point size or alpha.}

\item{annotate}{A \code{\link[=glue]{glue()}} string that is used to annotate the plot with additional properties of the height bin of the vp. Use \code{NULL} if no annotation is desired.}
}
\description{
Create a velocity azimuth display (VAD) plot
}
\examples{
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")
example_pvol <- read_pvolfile(pvolfile)
vp <- calculate_vp(pvolfile)
vad(example_pvol,
  range = c(5000, 30000), height = c(200, 400)
)
vad(example_pvol,
  vp = vp,
  height = 400,
  point_geom_args=list(ggplot2::aes(color=ZDR))
) +ggplot2::scale_color_gradient2()
vad(example_pvol,
  vp = vp, height=c(400,1200),
  point_geom=ggplot2::geom_bin2d,
  annotate="sdvvp: {round(sd_vvp,2)} [m/s]",
) + ggplot2::scale_fill_viridis_c()
}
