% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vad.R
\name{vad}
\alias{vad}
\alias{vad.pvol}
\title{Create a Velocity Azimuth Display (VAD) plot}
\usage{
vad(x, ...)

\method{vad}{pvol}(
  x,
  vp = NULL,
  ...,
  range = NULL,
  height = NULL,
  range_gate_filter = DBZH < 20 & RHOHV < 0.95,
  plotting_geom = ggplot2::geom_point,
  plotting_geom_args = list(),
  annotate = "{round(ff,1)} m/s, {round(dd)}°",
  annotation_size = 4,
  annotation_color = "red"
)
}
\arguments{
\item{x}{A polar volume from which range gates are extracted.}

\item{...}{Currently not used.}

\item{vp}{A vertical profile to annotate the VAD plot with the fit in the vertical profile}

\item{range}{The distance range in to filter the range gates with.
If a \code{vp} is provided the range is taken from there and the argument \code{range} should not be provided.}

\item{height}{The height range to filter the range gates by.
If only one value is provided next to a \code{vp} then the height bin intersection with this elevation is plotted.}

\item{range_gate_filter}{Optional filtering of the range gates. By default range gates are filtered for a \code{DBZH} less then 20 and a \code{RHOHV} less then 0.95.
Alternative filters could be used to highlight specific effects.}

\item{plotting_geom}{The geom function to visualize the range gates, the default is \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}, in some cases this suffers from over plotting.
Alternatives that avoid over plotting could be \code{\link[ggplot2:geom_bin_2d]{ggplot2::geom_bin2d()}} or \code{\link[ggpointdensity:geom_pointdensity]{ggpointdensity::geom_pointdensity()}}.}

\item{plotting_geom_args}{A list with additional arguments to the \code{plotting_geom} function. For example, controlling the point size or alpha.}

\item{annotate}{A \link[glue:glue]{glue} string that is used to annotate the plot with additional properties of the height bin of the \code{vp}.
The string is evaluated using the columns from \code{as.data.frame(vp)}.
Use \code{NULL} if no annotation is desired.}

\item{annotation_size}{The text size used for the annotation.}

\item{annotation_color}{The color used for the \code{vp} annotations and line.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
A velocity azimuth display plot visualized the radial velocity (\code{VRAD}) as a function of the azimuth from the radar.
Among others it can be used to asses the fit of the movement speeds in a vertical profile.
For example, when the movement is not uniform but rather in multiple directions or rotating this will be visible as
deviations from the sine function.
}
\details{
As a \link[ggplot2:ggplot]{ggplot2::ggplot} object  us returned additional elements can be added using regular function in
the \code{ggplot2} package. Labels could, for example, be modified using \code{\link[ggplot2:labs]{ggplot2::labs()}} or \code{\link[ggplot2:labs]{ggplot2::ggtitle()}}.
Using \code{\link[ggplot2:theme]{ggplot2::theme()}} the visual appearance can easily be modified. To do this the regular \link[ggplot2:gg-add]{+}
syntax can be used. In the examples this is demonstrated using scale.
}
\examples{
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")
example_pvol <- read_pvolfile(pvolfile)
# VAD plots can be created for polar volumes alone
vad(example_pvol,
  range = c(5000, 30000), height = c(200, 400)
)
# It is also possible to plot one height or more height bins from a `vp`.
# Many visual aspects can be controlled through the function arguments
# or through adding `ggplot`
vp <- calculate_vp(pvolfile)
vad(example_pvol,
  vp = vp,
  height = 400,
  plotting_geom_args = list(ggplot2::aes(color = ZDR))
) + ggplot2::scale_color_gradient2()
vad(example_pvol,
  vp = vp, height = c(400, 1200),
  plotting_geom = ggplot2::geom_bin2d,
  annotate = "sdvvp: {round(sd_vvp,2)} [m/s]",
) + ggplot2::scale_fill_viridis_c()
}
