% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{vol2bird}
\alias{vol2bird}
\title{Calculate a vertical profile of birds (VPB)}
\usage{
vol2bird(volume.in, profile.out = "", volume.out = "", verbose = F)
}
\arguments{
\item{volume.in}{A radar file containing a radar polar volume, either in
\href{http://www.eumetnet.eu/sites/default/files/OPERA2014_O4_ODIM_H5-v2.2.pdf}{ODIM}
format, which is the implementation of the OPERA data information model in \href{https://support.hdfgroup.org/HDF5/}{HDF5} format,
or a format supported by the \href{http://trmm-fc.gsfc.nasa.gov/trmm_gv/software/rsl/}{RSL library})}

\item{profile.out}{character string. Filename for the vertical profile to be generated in ODIM HDF5 format (optional)}

\item{volume.out}{character string. Filename for the polar volume to be generated in ODIM HDF5 format (optional, e.g. for converting RSL formats to ODIM)}

\item{verbose}{logical. When TRUE, pipe Docker stdout to R console}
}
\description{
Calculates a vertical profile of birds (VPB) from a polar volume
}
\details{
Requires a running \href{https://www.docker.com/}{Docker} daemon
}
\examples{
# locate example volume file:
volume <- system.file("extdata", "volume.h5", package="bioRad")
# copy to a home directory with read/write permissions:
file.copy(volume,"~/volume.h5")
# calculate the profile:
profile=vol2bird("~/volume.h5")
# plot the profile
plot(profile)
# clean up:
file.remove("~/volume.h5")
}

