% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{vol2bird_version}
\alias{vol2bird_version}
\title{Get installed vol2bird version}
\usage{
vol2bird_version(local_install)
}
\arguments{
\item{local_install}{Character. Path to local vol2bird installation, see
\code{\link[=calculate_vp]{calculate_vp()}} for details.}
}
\value{
\itemize{
\item Version number of installed vol2bird.
\item \code{NA}: Error: Docker system command not available.
\item \code{NaN}: Error: Docker daemon not running.
\item \code{NULL}: Error: Docker vol2bird image not available.
}
}
\description{
Returns the version of the vol2bird algorithm currently installed and used by
bioRad. This will either be the version installed in the Docker daemon
(default, see \code{\link[=check_docker]{check_docker()}}) or the version at \code{local_install}.
}
\examples{
\dontrun{
# Check installed vol2bird version
vol2bird_version()
}
}
\seealso{
\itemize{
\item \code{\link[=check_docker]{check_docker()}}
\item \code{\link[=update_docker]{update_docker()}}
}
}
