% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\docType{data}
\name{VP}
\alias{VP}
\title{Example object of class \code{VP} as generated by \code{\link{readVP}}}
\format{An object of class \code{VP} of length 4.}
\usage{
VP
}
\description{
Example object of class \code{VP} as generated by \code{\link{readVP}}
}
\keyword{datasets}

