% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{VPTimeSeries}
\alias{VPTimeSeries}
\title{Convert \code{VPList} to a single radar time series}
\usage{
VPTimeSeries(x, radar = NA)
}
\arguments{
\item{x}{An object of class \code{VPList}, usually a result of a call to \link[bioRad]{readVPList}}

\item{radar}{string containing the radar identifier to generate time series for. Only required when \code{VPList} object contains multiple radars}
}
\value{
an object of class \code{VP}, which is a list containing
\describe{
 \item{\code{VPList}}{vertical profiles, a list of \code{VP} objects, see \link[bioRad]{readVP} for a description}
 \item{\code{radar}}{string containing the radar identifier}
 \item{\code{dates}}{the \code{N} nominal times of the profiles}
 \item{\code{heights}}{the \code{M} heights of the layers in the profile}
 \item{\code{daterange}}{the minimum and maximum nominal time of the profiles in the list}
 \item{\code{timesteps}}{time differences between the profiles. Element \code{i} gives the time difference between profile \code{i} and \code{i+1}}
 \item{\code{data}}{list of \code{N} by \code{M} matrices containing the vertical profiles for each quantity measured}
 \item{\code{attributes}}{profile attributes, copied from the first profile contained in \code{x}}
 \item{\code{regular}}{logical indicating whether the time series is regular or not}
}
}
\description{
Convert \code{VPList} to a single radar time series
}
\examples{
\dontrun{readVP(c("my/path/profile1.h5","my/path/profile2.h5", ...))}

}

