% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{vpts}
\alias{vpts}
\title{Convert \code{vplist} to a single radar time series \code{vpts}}
\usage{
vpts(x, radar = NA)
}
\arguments{
\item{x}{An object of class \code{vplist}, usually a result of a call to \link[bioRad]{readvp.list}}

\item{radar}{string containing the radar identifier to generate time series for. Only required when \code{vplist} object contains multiple radars}
}
\value{
an object of class \code{vpts}, which is a list containing
\describe{
 \item{\code{radar}}{string containing the radar identifier}
 \item{\code{dates}}{the \code{N} nominal times of the profiles}
 \item{\code{heights}}{the \code{M} heights of the layers in the profile}
 \item{\code{daterange}}{the minimum and maximum nominal time of the profiles in the list}
 \item{\code{timesteps}}{time differences between the profiles. Element \code{i} gives the time difference between profile \code{i} and \code{i+1}}
 \item{\code{data}}{list of \code{N} by \code{M} matrices containing the vertical profiles for each quantity.
                    For a description of available quantities, see the \code{data} element of the \code{vp} class in \link[=summary.vp]{readvp}}
 \item{\code{attributes}}{profile attributes, copied from the first profile contained in \code{x}}
 \item{\code{regular}}{logical indicating whether the time series is regular or not}
}
}
\description{
Convert \code{vplist} to a single radar time series \code{vpts}
}
\examples{
\dontrun{readvp(c("my/path/profile1.h5","my/path/profile2.h5", ...))}

}

