% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioRad.R
\name{vpts}
\alias{vpts}
\title{Convert \code{vplist} to a single radar time series \code{vpts}}
\usage{
vpts(x, radar = NA)
}
\arguments{
\item{x}{An object of class \code{vplist}, usually a result of a call to \link[bioRad]{readvp.list}}

\item{radar}{string containing the radar identifier to generate time series for. Only required when \code{vplist} object contains multiple radars}
}
\value{
an object of class \link[=summary.vpts]{vpts}
}
\description{
Convert \code{vplist} to a single radar time series \code{vpts}
}
\examples{
\dontrun{readvp(c("my/path/profile1.h5","my/path/profile2.h5", ...))}

}
